Frozen Rush Data Storage

Freezes the Redis data base of an ArchiveAsyncTuning to a data.table::data.table(). No further points can be added to the archive but the data can be accessed and analyzed. Useful when the Redis data base is not permanently available. Use the callback mlr3tuning.async_freeze_archive to freeze the archive after the optimization has finished.

S3 Methods

as.data.table(archive)
ArchiveAsyncTuningFrozen -> data.table::data.table()
Returns a tabular view of all performed function calls of the Objective. The x_domain column is unnested to separate columns.

Super classes

bbotk::Archive -> bbotk::ArchiveAsync -> bbotk::ArchiveAsyncFrozen -> ArchiveAsyncTuningFrozen

Active bindings

internal_search_space

(paradox::ParamSet)
The search space containing those parameters that are internally optimized by the mlr3::Learner.

benchmark_result

(mlr3::BenchmarkResult)
Benchmark result.

Methods

Public methods

ArchiveAsyncTuningFrozen$new()

ArchiveAsyncTuningFrozen$learner()

ArchiveAsyncTuningFrozen$learners()

ArchiveAsyncTuningFrozen$learner_param_vals()

ArchiveAsyncTuningFrozen$predictions()

ArchiveAsyncTuningFrozen$resample_result()

ArchiveAsyncTuningFrozen$print()

ArchiveAsyncTuningFrozen$clone()

Inherited methods

bbotk::Archive$format()

bbotk::Archive$help()

bbotk::ArchiveAsync$best()

bbotk::ArchiveAsync$nds_selection()

bbotk::ArchiveAsyncFrozen$clear()

bbotk::ArchiveAsyncFrozen$data_with_state()

bbotk::ArchiveAsyncFrozen$pop_point()

bbotk::ArchiveAsyncFrozen$push_failed_point()

bbotk::ArchiveAsyncFrozen$push_points()

bbotk::ArchiveAsyncFrozen$push_result()

bbotk::ArchiveAsyncFrozen$push_running_point()

Method new()

Creates a new instance of this R6 class.

Usage

ArchiveAsyncTuningFrozen$new(archive)

Arguments

archive

(ArchiveAsyncTuning)
The archive to freeze.

Method learner()

Retrieve mlr3::Learner of the i-th evaluation, by position or by unique hash uhash. i and uhash are mutually exclusive. Learner does not contain a model. Use $learners() to get learners with models.

Usage

ArchiveAsyncTuningFrozen$learner(i = NULL, uhash = NULL)

Arguments

i

(integer(1))
The iteration value to filter for.

uhash

(logical(1))
The uhash value to filter for.

Method learners()

Retrieve list of trained mlr3::Learner objects of the i-th evaluation, by position or by unique hash uhash. i and uhash are mutually exclusive.

Usage

ArchiveAsyncTuningFrozen$learners(i = NULL, uhash = NULL)

Arguments

i

(integer(1))
The iteration value to filter for.

uhash

(logical(1))
The uhash value to filter for.

Method learner_param_vals()

Retrieve param values of the i-th evaluation, by position or by unique hash uhash. i and uhash are mutually exclusive.

Usage

ArchiveAsyncTuningFrozen$learner_param_vals(i = NULL, uhash = NULL)

Arguments

i

(integer(1))
The iteration value to filter for.

uhash

(logical(1))
The uhash value to filter for.

Method predictions()

Retrieve list of mlr3::Prediction objects of the i-th evaluation, by position or by unique hash uhash. i and uhash are mutually exclusive.

Usage

ArchiveAsyncTuningFrozen$predictions(i = NULL, uhash = NULL)

Arguments

i

(integer(1))
The iteration value to filter for.

uhash

(logical(1))
The uhash value to filter for.

Method resample_result()

Retrieve mlr3::ResampleResult of the i-th evaluation, by position or by unique hash uhash. i and uhash are mutually exclusive.

Usage

ArchiveAsyncTuningFrozen$resample_result(i = NULL, uhash = NULL)

Arguments

i

(integer(1))
The iteration value to filter for.

uhash

(logical(1))
The uhash value to filter for.

Method print()

Printer.

Usage

ArchiveAsyncTuningFrozen$print()

Arguments

...

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

ArchiveAsyncTuningFrozen$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.