Create Batch Tuning Callback

Specialized bbotk::CallbackBatch for batch tuning. Callbacks allow to customize the behavior of processes in mlr3tuning. The callback_batch_tuning() function creates a CallbackBatchTuning. Predefined callbacks are stored in the dictionary mlr_callbacks and can be retrieved with clbk(). For more information on tuning callbacks see callback_batch_tuning().

Super classes

mlr3misc::Callback -> bbotk::CallbackBatch -> CallbackBatchTuning

Public fields

on_eval_after_design

(function())
Stage called after design is created. Called in ObjectiveTuningBatch$eval_many().

on_eval_after_benchmark

(function())
Stage called after hyperparameter configurations are evaluated. Called in ObjectiveTuningBatch$eval_many().

on_eval_before_archive

(function())
Stage called before performance values are written to the archive. Called in ObjectiveTuningBatch$eval_many().

on_tuning_result_begin

(function())
Stage called before the results are written. Called in TuningInstance*$assign_result().

Methods

Public methods

CallbackBatchTuning$clone()

Inherited methods

mlr3misc::Callback$call()

mlr3misc::Callback$format()

mlr3misc::Callback$help()

mlr3misc::Callback$initialize()

mlr3misc::Callback$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

CallbackBatchTuning$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# write archive to disk
callback_batch_tuning("mlr3tuning.backup",
  on_optimization_end = function(callback, context) {
    saveRDS(context$instance$archive, "archive.rds")
  }
)
#> <CallbackBatchTuning:mlr3tuning.backup>
#> * Active Stages: on_optimization_end