FitnessFunction Class
Source: R/FitnessFunction.R

Implements a fitness function for mlr3 as R6 class FitnessFunction. An object of that class contains all relevant informations that are necessary to conduct tuning (mlr3::Task, mlr3::Learner, mlr3::Resampling, mlr3::Measures, paradox::ParamSet). After defining a fitness function, we can use it to predict the generalization error of a specific learner configuration defined by it's hyperparameter (using $eval()). The FitnessFunction class is the basis for further tuning strategies, i.e., grid or random search.

Usage

# Construction
ff = FitnessFunction$new(task, learner, resampling, param_set,
  ctrl = tune_control())
    # Public members
ff$task
ff$learner
ff$resampling
ff$param_set
ff$ctrl
ff$hooks
ff$bmr
    # Public methods
ff$eval(x)
ff$eval_vectorized(xts)
ff$get_best()
ff$run_hooks(id)

Arguments

task (mlr3::Task): The task that we want to evaluate.

learner (mlr3::Learner): The learner that we want to evaluate.

resampling (mlr3::Resampling): The Resampling method that is used to evaluate the learner.

param_set (paradox::ParamSet): Parameter set to define the hyperparameter space.

ctrl (list()): See tune_control().

xt (list()): A specific (transformed) parameter configuration given as named list (e.g. for rpart list(cp = 0.05, minsplit = 4)).

xts (list()): Collection of multiple (transformed) parameter values gained that is, for example, gained from a tuning strategy like grid search (see ?paradox::generate_design_grid).

id (character(1)): Identifier of a hook.

Details

$new() creates a new object of class FitnessFunction.

$task (mlr3::Task) the task for which the tuning should be conducted.

$learner (mlr3::Learner) the algorithm for which the tuning should be conducted.

$resampling (mlr3::Resampling) strategy to evaluate a parameter setting

$param_set (paradox::ParamSet) parameter space given to the Tuner object to generate parameter values.

$ctrl (list()) execution control object for tuning (see ?tune_control).

$hooks (list()) list of functions that could be executed with run_hooks().

$bmr (mlr3::BenchmarkResult) object that contains all tuning results as BenchmarkResult object (see ?BenchmarkResult).

$eval(xt) evaluates the (transformed) parameter setting xt (list) for the given learner and resampling.

$eval_vectorized(xts) performs resampling for multiple (transformed) parameter settings xts (list of lists).

$get_best() get best parameter configuration from the BenchmarkResult object.

$run_hooks() run a function that runs on the whole FitnessFunction object.

Examples

# Object required to define the fitness function:
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("holdout")
measures = mlr3::mlr_measures$mget("classif.ce")
task$measures = measures
param_set = paradox::ParamSet$new(params = list(
  paradox::ParamDbl$new("cp", lower = 0.001, upper = 0.1),
  paradox::ParamInt$new("minsplit", lower = 1, upper = 10)))

ff = FitnessFunction$new(
  task = task,
  learner = learner,
  resampling = resampling,
  param_set = param_set
)

ff$eval(data.table::data.table(cp = 0.05, minsplit = 5))

#> INFO [mlr3] Benchmarking 1 experiments
#> INFO [mlr3] Running learner 'classif.rpart1' on task 'iris' (iteration 1/1)'
#> INFO [mlr3] Finished benchmark

ff$eval(data.table::data.table(cp = 0.01, minsplit = 3))

#> INFO [mlr3] Benchmarking 1 experiments
#> INFO [mlr3] Running learner 'classif.rpart2' on task 'iris' (iteration 1/1)'
#> INFO [mlr3] Finished benchmark

ff$get_best()

#> <ResampleResult> of learner 'iris' on task 'classif.rpart1' with 1 iterations
#>     Measure Min. 1st Qu. Median Mean 3rd Qu. Max. Sd
#>  classif.ce 0.08    0.08   0.08 0.08    0.08 0.08 NA