Class for Tuning Objective

Stores the objective function that estimates the performance of hyperparameter configurations. This class is usually constructed internally by the TuningInstanceSingleCrit or TuningInstanceMultiCrit.

Super class

bbotk::Objective -> ObjectiveTuning

Public fields

task

(mlr3::Task).

learner

(mlr3::Learner).

resampling

(mlr3::Resampling).

measures

(list of mlr3::Measure).

store_models

(logical(1)).

store_benchmark_result

(logical(1)).

archive

(ArchiveTuning).

hotstart_stack

(mlr3::HotstartStack).

allow_hotstart

(logical(1)).

keep_hotstart_stack

(logical(1)).

callbacks

(List of CallbackTunings).

Methods

Public methods

ObjectiveTuning$new()

ObjectiveTuning$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$eval_dt()

bbotk::Objective$eval_many()

bbotk::Objective$format()

bbotk::Objective$print()

Method new()

Creates a new instance of this R6 class.

Usage

ObjectiveTuning$new(
  task,
  learner,
  resampling,
  measures,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = TRUE,
  allow_hotstart = FALSE,
  keep_hotstart_stack = FALSE,
  archive = NULL,
  callbacks = list()
)

Arguments

task

(mlr3::Task)
Task to operate on.

learner

(mlr3::Learner)
Learner to tune.

resampling

(mlr3::Resampling)
Resampling that is used to evaluated the performance of the hyperparameter configurations. Uninstantiated resamplings are instantiated during construction so that all configurations are evaluated on the same data splits. Already instantiated resamplings are kept unchanged. Specialized Tuner change the resampling e.g. to evaluate a hyperparameter configuration on different data splits. This field, however, always returns the resampling passed in construction.

measures

(list of mlr3::Measure)
Measures to optimize.

store_benchmark_result

(logical(1))
If TRUE (default), store resample result of evaluated hyperparameter configurations in archive as mlr3::BenchmarkResult.

store_models

(logical(1))
If TRUE, fitted models are stored in the benchmark result (archive$benchmark_result). If store_benchmark_result = FALSE, models are only stored temporarily and not accessible after the tuning. This combination is needed for measures that require a model.

check_values

(logical(1))
If TRUE, hyperparameter values are checked before evaluation and performance scores after. If FALSE (default), values are unchecked but computational overhead is reduced.

allow_hotstart

(logical(1))
Allow to hotstart learners with previously fitted models. See also mlr3::HotstartStack. The learner must support hotstarting. Sets store_models = TRUE.

keep_hotstart_stack

(logical(1))
If TRUE, mlr3::HotstartStack is kept in $objective$hotstart_stack after tuning.

archive

(ArchiveTuning)
Reference to archive of TuningInstanceSingleCrit | TuningInstanceMultiCrit. If NULL (default), benchmark result and models cannot be stored.

callbacks

(list of CallbackTuning)
List of callbacks.

Method clone()

The objects of this class are cloneable with this method.

Usage

ObjectiveTuning$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.