Class for Tuning Objective

Stores the objective function that estimates the performance of hyperparameter configurations. This class is usually constructed internally by the TuningInstanceBatchSingleCrit or TuningInstanceBatchMultiCrit.

Super classes

bbotk::Objective -> mlr3tuning::ObjectiveTuning -> ObjectiveTuningAsync

Methods

Public methods

ObjectiveTuningAsync$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$eval_dt()

bbotk::Objective$eval_many()

bbotk::Objective$format()

bbotk::Objective$help()

bbotk::Objective$print()

mlr3tuning::ObjectiveTuning$initialize()

Method clone()

The objects of this class are cloneable with this method.

Usage

ObjectiveTuningAsync$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.