Class for Tuning Objective

Stores the objective function that estimates the performance of hyperparameter configurations. This class is usually constructed internally by the TuningInstanceBatchSingleCrit or TuningInstanceBatchMultiCrit.

Super classes

bbotk::Objective -> mlr3tuning::ObjectiveTuning -> ObjectiveTuningBatch

Public fields

archive

(ArchiveBatchTuning).

Methods

Public methods

ObjectiveTuningBatch$new()

ObjectiveTuningBatch$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$eval_dt()

bbotk::Objective$eval_many()

bbotk::Objective$format()

bbotk::Objective$print()

Method new()

Creates a new instance of this R6 class.

Usage

ObjectiveTuningBatch$new(
  task,
  learner,
  resampling,
  measures,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE,
  archive = NULL,
  callbacks = NULL
)

Arguments

task

(mlr3::Task)
Task to operate on.

learner

(mlr3::Learner)
Learner to tune.

resampling

(mlr3::Resampling)
Resampling that is used to evaluate the performance of the hyperparameter configurations. Uninstantiated resamplings are instantiated during construction so that all configurations are evaluated on the same data splits. Already instantiated resamplings are kept unchanged. Specialized Tuner change the resampling e.g. to evaluate a hyperparameter configuration on different data splits. This field, however, always returns the resampling passed in construction.

measures

(list of mlr3::Measure)
Measures to optimize.

store_benchmark_result

(logical(1))
If TRUE (default), store resample result of evaluated hyperparameter configurations in archive as mlr3::BenchmarkResult.

store_models

(logical(1))
If TRUE, fitted models are stored in the benchmark result (archive$benchmark_result). If store_benchmark_result = FALSE, models are only stored temporarily and not accessible after the tuning. This combination is needed for measures that require a model.

check_values

(logical(1))
If TRUE, hyperparameter values are checked before evaluation and performance scores after. If FALSE (default), values are unchecked but computational overhead is reduced.

archive

(ArchiveBatchTuning)
Reference to archive of TuningInstanceBatchSingleCrit | TuningInstanceBatchMultiCrit. If NULL (default), benchmark result and models cannot be stored.

callbacks

(list of Callback)
List of callbacks.

Method clone()

The objects of this class are cloneable with this method.

Usage

ObjectiveTuningBatch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.