Abstract Tuner Class
Source: R/Tuner.R

Abstract Tuner class that implements the main functionality each tuner must have. A tuner is an object that describes the tuning strategy how to search the hyperparameter space given within the [FitnessFunction] object.

Usage

# Construction
tuner = Tuner$new(id, ff, terminator, settings = list())
    # public members
tuner$id
tuner$ff
tuner$terminator
tuner$settings
    # public methods
tuner$tune()
tuner$tune_result()
tuner$aggregated(unnest)

Arguments

id (character(1)):
The id of the Tuner.

ff ([FitnessFunction]).

terminator ([Terminator]).

settings (list):
The settings for the Tuner.

unnest (logical(1)):
If TRUE returns each parameter as a column of the data table. Otherwise the column includes a list of the parameter.

Details

$new() creates a new object of class [Tuner].

id stores an identifier for this [Tuner].

ff stores the FitnessFunction to optimize.

terminator stores the [Terminator].

settings is a list of hyperparamter settings for this [Tuner].

tune() performs the tuning, until the budget of the [Terminator] in the [FitnessFunction] is exhausted.

tune_result() returns a list with 2 elements:

performance (numeric()) with the best performance.

param_set$values (numeric()) with corresponding hyperparameters.

See also