Class for Asynchronous Tuning Algorithms

The TunerAsync implements the asynchronous optimization algorithm.

Details

TunerAsync is an abstract base class that implements the base functionality each asynchronous tuner must provide.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Getting started with hyperparameter optimization.

An overview of all tuners can be found on our website.

Tune a support vector machine on the Sonar data set.

Learn about tuning spaces.

Estimate the model performance with nested resampling.

Learn about multi-objective optimization.

Simultaneously optimize hyperparameters and use early stopping with XGBoost.

Automate the tuning.

The gallery features a collection of case studies and demos about optimization.

Learn more advanced methods with the Practical Tuning Series.

Learn about hotstarting models.

Run the default hyperparameter configuration of learners as a baseline.

Use the Hyperband optimizer with different budget parameters.

The cheatsheet summarizes the most important functions of mlr3tuning.

Super class

mlr3tuning::Tuner -> TunerAsync

Methods

Public methods

TunerAsync$optimize()

TunerAsync$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$initialize()

mlr3tuning::Tuner$print()

Method optimize()

Performs the tuning on a TuningInstanceAsyncSingleCrit or TuningInstanceAsyncMultiCrit until termination. The single evaluations will be written into the ArchiveAsyncTuning that resides in the TuningInstanceAsyncSingleCrit/TuningInstanceAsyncMultiCrit. The result will be written into the instance object.

Usage

TunerAsync$optimize(inst)

Arguments

inst

(TuningInstanceAsyncSingleCrit | TuningInstanceAsyncMultiCrit).

Returns

data.table::data.table()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerAsync$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.