TunerAsyncFromOptimizerAsync

Internally used to transform bbotk::Optimizer to Tuner.

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerAsync -> TunerAsyncFromOptimizerAsync

Active bindings

param_set

(paradox::ParamSet)
Set of control parameters.

Methods

Public methods

TunerAsyncFromOptimizerAsync$new()

TunerAsyncFromOptimizerAsync$optimize()

TunerAsyncFromOptimizerAsync$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerAsyncFromOptimizerAsync$new(optimizer, man = NA_character_)

Arguments

optimizer

bbotk::Optimizer
Optimizer that is called.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method optimize()

Performs the tuning on a TuningInstanceBatchSingleCrit / TuningInstanceBatchMultiCrit until termination. The single evaluations and the final results will be written into the ArchiveAsyncTuning that resides in the TuningInstanceBatchSingleCrit/TuningInstanceBatchMultiCrit. The final result is returned.

Usage

TunerAsyncFromOptimizerAsync$optimize(inst)

Arguments

inst

(TuningInstanceBatchSingleCrit | TuningInstanceBatchMultiCrit).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerAsyncFromOptimizerAsync$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.