TunerGridSearch
Source: R/TunerGridSearch.R

Tuner child class to conduct grid search.

Usage

tuner = TunerGridSearch$new(ff, terminator, resolution)

See Tuner for a description of the interface.

Arguments

ff (FitnessFunction): Black box function used for the tuning.

terminator (Terminator): Terminator to control the termination. Will be triggered by the hooks of the FitnessFunction.

resolution (integer(1)): Resolution of the grid. If none is specified we will try to calculate the resolution form the Terminator.

Details

$new() creates a new object of class TunerGridSearch. The interface is described in Tuner.

See also

Examples

task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv")
resampling$param_set$values$folds = 2
measures = mlr3::mlr_measures$mget("classif.ce")
task$measures = measures
param_set = paradox::ParamSet$new(
  params = list(
    paradox::ParamDbl$new("cp", lower = 0.001, upper = 0.1)
  )
)
ff = FitnessFunction$new(task, learner, resampling, param_set)

terminator = TerminatorEvaluations$new(10)
gs = TunerGridSearch$new(ff, terminator)
gs$tune()$tune_result()

#> INFO [mlr3] Benchmarking 20 experiments
#> INFO [mlr3] Running learner 'classif.rpart1' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart1' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart2' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart2' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart3' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart3' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart4' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart4' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart5' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart5' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart6' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart6' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart7' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart7' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart8' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart8' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart9' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart9' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Running learner 'classif.rpart10' on task 'iris' (iteration 1/2)'
#> INFO [mlr3] Running learner 'classif.rpart10' on task 'iris' (iteration 2/2)'
#> INFO [mlr3] Finished benchmark

#> $performance
#> classif.ce 
#> 0.04666667 
#> 
#> $values
#> $values$cp
#> [1] 0.1
#> 
#>