Class for Multi Criteria Tuning

The TuningInstanceBatchMultiCrit specifies a tuning problem for a Tuner. The function ti() creates a TuningInstanceBatchMultiCrit and the function tune() creates an instance internally.

Details

The instance contains an ObjectiveTuningBatch object that encodes the black box objective function a Tuner has to optimize. The instance allows the basic operations of querying the objective at design points ($eval_batch()). This operation is usually done by the Tuner. Evaluations of hyperparameter configurations are performed in batches by calling mlr3::benchmark() internally. The evaluated hyperparameter configurations are stored in the ArchiveBatchTuning ($archive). Before a batch is evaluated, the bbotk::Terminator is queried for the remaining budget. If the available budget is exhausted, an exception is raised, and no further evaluations can be performed from this point on. The tuner is also supposed to store its final result, consisting of a selected hyperparameter configuration and associated estimated performance values, by calling the method instance$assign_result.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Learn about multi-objective optimization.

The gallery features a collection of case studies and demos about optimization.

Analysis

For analyzing the tuning results, it is recommended to pass the ArchiveBatchTuning to as.data.table(). The returned data table is joined with the benchmark result which adds the mlr3::ResampleResult for each hyperparameter evaluation.

The archive provides various getters (e.g. $learners()) to ease the access. All getters extract by position (i) or unique hash (uhash). For a complete list of all getters see the methods section.

The benchmark result ($benchmark_result) allows to score the hyperparameter configurations again on a different measure. Alternatively, measures can be supplied to as.data.table().

The mlr3viz package provides visualizations for tuning results.

Super classes

bbotk::OptimInstance -> bbotk::OptimInstanceBatch -> bbotk::OptimInstanceBatchMultiCrit -> TuningInstanceBatchMultiCrit

Active bindings

result_learner_param_vals

(list())
List of param values for the optimal learner call.

Methods

Public methods

TuningInstanceBatchMultiCrit$new()

TuningInstanceBatchMultiCrit$assign_result()

TuningInstanceBatchMultiCrit$clone()

Inherited methods

bbotk::OptimInstance$clear()

bbotk::OptimInstance$format()

bbotk::OptimInstance$print()

bbotk::OptimInstanceBatch$eval_batch()

bbotk::OptimInstanceBatch$objective_function()

Method new()

Creates a new instance of this R6 class.

Usage

TuningInstanceBatchMultiCrit$new(
  task,
  learner,
  resampling,
  measures,
  terminator,
  search_space = NULL,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE,
  callbacks = NULL
)

Arguments

task

(mlr3::Task)
Task to operate on.

learner

(mlr3::Learner)
Learner to tune.

resampling

(mlr3::Resampling)
Resampling that is used to evaluate the performance of the hyperparameter configurations. Uninstantiated resamplings are instantiated during construction so that all configurations are evaluated on the same data splits. Already instantiated resamplings are kept unchanged. Specialized Tuner change the resampling e.g. to evaluate a hyperparameter configuration on different data splits. This field, however, always returns the resampling passed in construction.

measures

(list of mlr3::Measure)
Measures to optimize.

terminator

(Terminator)
Stop criterion of the tuning process.

search_space

(paradox::ParamSet)
Hyperparameter search space. If NULL (default), the search space is constructed from the TuneToken of the learner's parameter set (learner$param_set).

store_benchmark_result

(logical(1))
If TRUE (default), store resample result of evaluated hyperparameter configurations in archive as mlr3::BenchmarkResult.

store_models

(logical(1))
If TRUE, fitted models are stored in the benchmark result (archive$benchmark_result). If store_benchmark_result = FALSE, models are only stored temporarily and not accessible after the tuning. This combination is needed for measures that require a model.

check_values

(logical(1))
If TRUE, hyperparameter values are checked before evaluation and performance scores after. If FALSE (default), values are unchecked but computational overhead is reduced.

callbacks

(list of Callback)
List of callbacks.

Method assign_result()

The Tuner object writes the best found points and estimated performance values here. For internal use.

Usage

TuningInstanceBatchMultiCrit$assign_result(xdt, ydt, learner_param_vals = NULL)

Arguments

xdt

(data.table::data.table())
Hyperparameter values as data.table::data.table(). Each row is one configuration. Contains values in the search space. Can contain additional columns for extra information.

ydt

(data.table::data.table())
Optimal outcomes, e.g. the Pareto front.

learner_param_vals

(List of named list()s)
Fixed parameter values of the learner that are neither part of the

Method clone()

The objects of this class are cloneable with this method.

Usage

TuningInstanceBatchMultiCrit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter optimization on the Palmer Penguins data set
task = tsk("penguins")

# Load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# Construct tuning instance
instance = ti(
  task = task,
  learner = learner,
  resampling = rsmp("cv", folds = 3),
  measures = msrs(c("classif.ce", "time_train")),
  terminator = trm("evals", n_evals = 4)
)

# Choose optimization algorithm
tuner = tnr("random_search", batch_size = 2)

# Run tuning
tuner$optimize(instance)
#>           cp learner_param_vals  x_domain classif.ce  time_train
#>        <num>             <list>    <list>      <num>       <num>
#> 1: -7.349764          <list[2]> <list[1]> 0.06402237 0.004666667

# Optimal hyperparameter configurations
instance$result
#>           cp learner_param_vals  x_domain classif.ce  time_train
#>        <num>             <list>    <list>      <num>       <num>
#> 1: -7.349764          <list[2]> <list[1]> 0.06402237 0.004666667

# Inspect all evaluated configurations
as.data.table(instance$archive)
#>           cp classif.ce  time_train  x_domain_cp runtime_learners
#>        <num>      <num>       <num>        <num>            <num>
#> 1: -7.095299 0.06402237 0.005000000 0.0008289927            0.029
#> 2: -7.349764 0.06402237 0.004666667 0.0006427440            0.027
#> 3: -2.357433 0.06692093 0.005000000 0.0946629107            0.028
#> 4: -8.488297 0.06402237 0.010333333 0.0002058636            0.046
#>              timestamp batch_nr warnings errors  resample_result
#>                 <POSc>    <int>    <int>  <int>           <list>
#> 1: 2024-05-04 07:55:18        1        0      0 <ResampleResult>
#> 2: 2024-05-04 07:55:18        1        0      0 <ResampleResult>
#> 3: 2024-05-04 07:55:18        2        0      0 <ResampleResult>
#> 4: 2024-05-04 07:55:18        2        0      0 <ResampleResult>