Single Criterion Tuning Instance

Specifies a general single-criteria tuning scenario, including objective function and archive for Tuners to act upon. This class stores an ObjectiveTuning object that encodes the black box objective function which a Tuner has to optimize. It allows the basic operations of querying the objective at design points ($eval_batch()), storing the evaluations in the internal Archive and accessing the final result ($result).

Evaluations of hyperparameter configurations are performed in batches by calling mlr3::benchmark() internally. Before a batch is evaluated, the bbotk::Terminator is queried for the remaining budget. If the available budget is exhausted, an exception is raised, and no further evaluations can be performed from this point on.

The tuner is also supposed to store its final result, consisting of a selected hyperparameter configuration and associated estimated performance values, by calling the method instance$assign_result.

Super classes

bbotk::OptimInstance -> bbotk::OptimInstanceSingleCrit -> TuningInstanceSingleCrit

Active bindings

result_learner_param_vals

(list())
Param values for the optimal learner call.

Methods

Public methods

TuningInstanceSingleCrit$new()

TuningInstanceSingleCrit$assign_result()

TuningInstanceSingleCrit$clone()

Inherited methods

bbotk::OptimInstance$eval_batch()

bbotk::OptimInstance$format()

bbotk::OptimInstance$objective_function()

bbotk::OptimInstance$print()

Method new()

Creates a new instance of this R6 class.

This defines the resampled performance of a learner on a task, a feasibility region for the parameters the tuner is supposed to optimize, and a termination criterion.

Usage

TuningInstanceSingleCrit$new(
  task,
  learner,
  resampling,
  measure,
  terminator,
  search_space = NULL,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE
)

Arguments

task

(mlr3::Task)
Task to operate on.

learner

(mlr3::Learner).

resampling

(mlr3::Resampling)
Uninstantiated resamplings are instantiated during construction so that all configurations are evaluated on the same data splits.

measure

(mlr3::Measure)
Measure to optimize.

terminator

(Terminator).

search_space

(paradox::ParamSet).

store_benchmark_result

(logical(1))
Store benchmark result in archive?

store_models

(logical(1))
Store models in benchmark result?

check_values

(logical(1))
Should parameters before the evaluation and the results be checked for validity?

Method assign_result()

The Tuner object writes the best found point and estimated performance value here. For internal use.

Usage

TuningInstanceSingleCrit$assign_result(xdt, y, learner_param_vals = NULL)

Arguments

xdt

(data.table::data.table())
x values as data.table. Each row is one point. Contains the value in the search space of the TuningInstanceMultiCrit object. Can contain additional columns for extra information.

y

(numeric(1))
Optimal outcome.

learner_param_vals

(list())
Fixed parameter values of the learner that are neither part of the

Method clone()

The objects of this class are cloneable with this method.

Usage

TuningInstanceSingleCrit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)
library(paradox)
library(mlr3)

# Objects required to define the performance evaluator:
task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measure = msr("classif.ce")
param_set = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1),
  ParamInt$new("minsplit", lower = 1, upper = 10))
)

terminator = trm("evals", n_evals = 5)
inst = TuningInstanceSingleCrit$new(
  task = task,
  learner = learner,
  resampling = resampling,
  measure = measure,
  search_space = param_set,
  terminator = terminator
)

# first 4 points as cross product
design = CJ(cp = c(0.05, 0.01), minsplit = c(5, 3))
inst$eval_batch(design)
inst$archive

#> <ArchiveTuning>
#>      cp minsplit classif.ce                                uhash  x_domain
#> 1: 0.01        3       0.02 e2f42929-40c1-414f-afdb-3295facdcd27 <list[2]>
#> 2: 0.01        5       0.02 a936c143-17b9-4fd4-8e73-91d3e72a2d63 <list[2]>
#> 3: 0.05        3       0.06 62c9c32f-2cec-42ee-84af-c23df909bcf6 <list[2]>
#> 4: 0.05        5       0.06 647e201f-422b-4094-83a5-483ab2041b9c <list[2]>
#>              timestamp batch_nr
#> 1: 2020-12-31 04:52:54        1
#> 2: 2020-12-31 04:52:54        1
#> 3: 2020-12-31 04:52:54        1
#> 4: 2020-12-31 04:52:54        1

# try more points, catch the raised terminated message
tryCatch(
  inst$eval_batch(data.table(cp = 0.01, minsplit = 7)),
  terminated_error = function(e) message(as.character(e))
)

# try another point although the budget is now exhausted
# -> no extra evaluations
tryCatch(
  inst$eval_batch(data.table(cp = 0.01, minsplit = 9)),
  terminated_error = function(e) message(as.character(e))
)

#> Error: Objective (obj:classif.rpart_on_iris, term:<TerminatorEvals>) terminated

inst$archive

#> <ArchiveTuning>
#>      cp minsplit classif.ce                                uhash  x_domain
#> 1: 0.01        3       0.02 e2f42929-40c1-414f-afdb-3295facdcd27 <list[2]>
#> 2: 0.01        5       0.02 a936c143-17b9-4fd4-8e73-91d3e72a2d63 <list[2]>
#> 3: 0.05        3       0.06 62c9c32f-2cec-42ee-84af-c23df909bcf6 <list[2]>
#> 4: 0.05        5       0.06 647e201f-422b-4094-83a5-483ab2041b9c <list[2]>
#> 5: 0.01        7       0.02 eb3ce032-ee25-48e4-8b36-6cb58afe2000 <list[2]>
#>              timestamp batch_nr
#> 1: 2020-12-31 04:52:54        1
#> 2: 2020-12-31 04:52:54        1
#> 3: 2020-12-31 04:52:54        1
#> 4: 2020-12-31 04:52:54        1
#> 5: 2020-12-31 04:52:54        2

### Error handling
# get a learner which breaks with 50% probability
# set encapsulation + fallback
learner = lrn("classif.debug", error_train = 0.5)
learner$encapsulate = c(train = "evaluate", predict = "evaluate")
learner$fallback = lrn("classif.featureless")

param_set = ParamSet$new(list(
  ParamDbl$new("x", lower = 0, upper = 1)
))

inst = TuningInstanceSingleCrit$new(
  task = tsk("wine"),
  learner = learner,
  resampling = rsmp("cv", folds = 3),
  measure = msr("classif.ce"),
  search_space = param_set,
  terminator = trm("evals", n_evals = 5)
)

tryCatch(
  inst$eval_batch(data.table(x = 1:5 / 5)),
  terminated_error = function(e) message(as.character(e))
)

archive = as.data.table(inst$archive)

# column errors: multiple errors recorded
print(archive)

#>      x classif.ce                                uhash           timestamp
#> 1: 0.2  0.6908663 514c1fac-cf2a-47ff-8b43-52469e2088f4 2020-12-31 04:52:54
#> 2: 0.4  0.6011299 0df6ce88-4c64-465b-a43c-bb2c6c8d5aa3 2020-12-31 04:52:54
#> 3: 0.6  0.6011299 f8da08b3-409d-4b44-86de-534c792447d4 2020-12-31 04:52:54
#> 4: 0.8  0.6011299 2bf82b11-e21f-46e7-a93c-6b6566913f58 2020-12-31 04:52:54
#> 5: 1.0  0.6739171 1ae64177-2ad1-4331-a58b-2ebe4836f752 2020-12-31 04:52:54
#>    batch_nr x_domain_x
#> 1:        1        0.2
#> 2:        1        0.4
#> 3:        1        0.6
#> 4:        1        0.8
#> 5:        1        1.0