Create Asynchronous Tuning Callback

Function to create a CallbackAsyncTuning. Predefined callbacks are stored in the dictionary mlr_callbacks and can be retrieved with clbk().

Tuning callbacks can be called from different stages of the tuning process. The stages are prefixed with on_*.

Start Tuning
     - on_optimization_begin
    Start Worker
         - on_worker_begin
        Start Evaluation
             - on_eval_after_xs
             - on_eval_after_resample
             - on_eval_before_archive
        End Evaluation
         - on_worker_end
    End Worker
     - on_result
     - on_optimization_end
End Tuning

See also the section on parameters for more information on the stages. A tuning callback works with ContextAsyncTuning.

Usage

callback_async_tuning(
  id,
  label = NA_character_,
  man = NA_character_,
  on_optimization_begin = NULL,
  on_worker_begin = NULL,
  on_eval_after_xs = NULL,
  on_eval_after_resample = NULL,
  on_eval_before_archive = NULL,
  on_worker_end = NULL,
  on_result = NULL,
  on_optimization_end = NULL
)

Arguments

id

(character(1))
Identifier for the new instance.

label

(character(1))
Label for the new instance.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

on_optimization_begin

(function())
Stage called at the beginning of the optimization. Called in Optimizer$optimize().

on_worker_begin

(function())
Stage called at the beginning of the optimization on the worker. Called in the worker loop.

on_eval_after_xs

(function())
Stage called after xs is passed. Called in ObjectiveTuning$eval().

on_eval_after_resample

(function())
Stage called after a hyperparameter configuration is evaluated. Called in ObjectiveTuning$eval().

on_eval_before_archive

(function())
Stage called before performance values are written to the archive. Called in ObjectiveTuning$eval().

on_worker_end

(function())
Stage called at the end of the optimization on the worker. Called in the worker loop.

on_result

(function())
Stage called after the result is written. Called in OptimInstance$assign_result().

on_optimization_end

(function())
Stage called at the end of the optimization. Called in Optimizer$optimize().

Details

When implementing a callback, each function must have two arguments named callback and context. A callback can write data to the state ($state), e.g. settings that affect the callback itself. Tuning callbacks access ContextAsyncTuning.