Hyperparameter Tuning with Generalized Simulated Annealing

Subclass for generalized simulated annealing tuning. Calls GenSA::GenSA() from package GenSA.

Source

Tsallis C, Stariolo DA (1996). “Generalized simulated annealing.” Physica A: Statistical Mechanics and its Applications, 233(1-2), 395–406. doi:10.1016/s0378-4371(96)00271-3 .

Xiang Y, Gubian S, Suomela B, Hoeng J (2013). “Generalized Simulated Annealing for Global Optimization: The GenSA Package.” The R Journal, 5(1), 13. doi:10.32614/rj-2013-002 .

Details

In contrast to the GenSA::GenSA() defaults, we set smooth = FALSE as a default.

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("gensa")

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria. A batch contains of batch_size times resampling$iters jobs. E.g., if you set a batch size of 10 points and do a 5-fold cross validation, you can utilize up to 50 cores.

Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerBatchGenSA which can be applied on any black box optimization problem. See also the documentation of bbotk.

Parameters

smooth

logical(1)

temperature

numeric(1)

acceptance.param

numeric(1)

verbose

logical(1)

trace.mat

logical(1)

For the meaning of the control parameters, see GenSA::GenSA(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

In contrast to the GenSA::GenSA() defaults, we set trace.mat = FALSE. Note that GenSA::GenSA() uses smooth = TRUE as a default. In the case of using this optimizer for Hyperparameter Optimization you may want to set smooth = FALSE.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Getting started with hyperparameter optimization.

An overview of all tuners can be found on our website.

Tune a support vector machine on the Sonar data set.

Learn about tuning spaces.

Estimate the model performance with nested resampling.

Learn about multi-objective optimization.

Simultaneously optimize hyperparameters and use early stopping with XGBoost.

Automate the tuning.

The gallery features a collection of case studies and demos about optimization.

Learn more advanced methods with the Practical Tuning Series.

Learn about hotstarting models.

Run the default hyperparameter configuration of learners as a baseline.

Use the Hyperband optimizer with different budget parameters.

The cheatsheet summarizes the most important functions of mlr3tuning.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerBatch -> mlr3tuning::TunerBatchFromOptimizerBatch -> TunerBatchGenSA

Methods

Public methods

TunerBatchGenSA$new()

TunerBatchGenSA$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerBatchFromOptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerBatchGenSA$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerBatchGenSA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  tuner = tnr("gensa"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10
)
#> Warning: one-dimensional optimization by Nelder-Mead is unreliable:
#> use "Brent" or optimize() directly

# best performing hyperparameter configuration
instance$result
#>           cp learner_param_vals  x_domain classif.ce
#>        <num>             <list>    <list>      <num>
#> 1: -5.721042          <list[2]> <list[1]> 0.04347826

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp classif.ce runtime_learners           timestamp warnings errors
#>         <num>      <num>            <num>              <POSc>    <int>  <int>
#>  1: -5.721042 0.04347826            0.006 2024-11-25 08:44:29        0      0
#>  2: -2.850714 0.06086957            0.005 2024-11-25 08:44:29        0      0
#>  3: -7.568995 0.04347826            0.005 2024-11-25 08:44:29        0      0
#>  4: -5.721042 0.04347826            0.006 2024-11-25 08:44:29        0      0
#>  5: -5.721042 0.04347826            0.005 2024-11-25 08:44:29        0      0
#>  6: -5.721042 0.04347826            0.005 2024-11-25 08:44:29        0      0
#>  7: -5.148938 0.04347826            0.006 2024-11-25 08:44:30        0      0
#>  8: -6.293146 0.04347826            0.005 2024-11-25 08:44:30        0      0
#>  9: -6.007094 0.04347826            0.005 2024-11-25 08:44:30        0      0
#> 10: -5.434990 0.04347826            0.006 2024-11-25 08:44:30        0      0
#>      x_domain batch_nr  resample_result
#>        <list>    <int>           <list>
#>  1: <list[1]>        1 <ResampleResult>
#>  2: <list[1]>        2 <ResampleResult>
#>  3: <list[1]>        3 <ResampleResult>
#>  4: <list[1]>        4 <ResampleResult>
#>  5: <list[1]>        5 <ResampleResult>
#>  6: <list[1]>        6 <ResampleResult>
#>  7: <list[1]>        7 <ResampleResult>
#>  8: <list[1]>        8 <ResampleResult>
#>  9: <list[1]>        9 <ResampleResult>
#> 10: <list[1]>       10 <ResampleResult>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))