Hyperparameter Tuning with Grid Search

Subclass for grid search tuning.

Details

The grid is constructed as a Cartesian product over discretized values per parameter, see paradox::generate_design_grid(). If the learner supports hotstarting, the grid is sorted by the hotstart parameter (see also mlr3::HotstartStack). If not, the points of the grid are evaluated in a random order.

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("grid_search")

Control Parameters

resolution

integer(1)
Resolution of the grid, see paradox::generate_design_grid().

param_resolutions

named integer()
Resolution per parameter, named by parameter ID, see paradox::generate_design_grid().

batch_size

integer(1)
Maximum number of points to try in a batch.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria. A batch contains of batch_size times resampling$iters jobs. E.g., if you set a batch size of 10 points and do a 5-fold cross validation, you can utilize up to 50 cores.

Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerBatchGridSearch which can be applied on any black box optimization problem. See also the documentation of bbotk.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

An overview of all tuners can be found on our website.

Learn more about tuners.

The gallery features a collection of case studies and demos about optimization.

Use the Hyperband optimizer with different budget parameters.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerBatch -> mlr3tuning::TunerBatchFromOptimizerBatch -> TunerBatchGridSearch

Methods

Public methods

TunerBatchGridSearch$new()

TunerBatchGridSearch$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerBatchFromOptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerBatchGridSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerBatchGridSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  tuner = tnr("grid_search"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10
)

# best performing hyperparameter configuration
instance$result
#>           cp learner_param_vals  x_domain classif.ce
#>        <num>             <list>    <list>      <num>
#> 1: -2.302585          <list[2]> <list[1]> 0.04347826

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp classif.ce  x_domain_cp runtime_learners           timestamp
#>         <num>      <num>        <num>            <num>              <POSc>
#>  1: -2.302585 0.04347826 0.1000000000            0.011 2024-06-28 13:09:46
#>  2: -4.605170 0.04347826 0.0100000000            0.011 2024-06-28 13:09:46
#>  3: -3.070113 0.04347826 0.0464158883            0.011 2024-06-28 13:09:46
#>  4: -3.837642 0.04347826 0.0215443469            0.012 2024-06-28 13:09:46
#>  5: -6.907755 0.04347826 0.0010000000            0.011 2024-06-28 13:09:46
#>  6: -8.442812 0.04347826 0.0002154435            0.011 2024-06-28 13:09:46
#>  7: -7.675284 0.04347826 0.0004641589            0.010 2024-06-28 13:09:46
#>  8: -9.210340 0.04347826 0.0001000000            0.013 2024-06-28 13:09:46
#>  9: -6.140227 0.04347826 0.0021544347            0.011 2024-06-28 13:09:46
#> 10: -5.372699 0.04347826 0.0046415888            0.011 2024-06-28 13:09:46
#>     batch_nr warnings errors  resample_result
#>        <int>    <int>  <int>           <list>
#>  1:        1        0      0 <ResampleResult>
#>  2:        2        0      0 <ResampleResult>
#>  3:        3        0      0 <ResampleResult>
#>  4:        4        0      0 <ResampleResult>
#>  5:        5        0      0 <ResampleResult>
#>  6:        6        0      0 <ResampleResult>
#>  7:        7        0      0 <ResampleResult>
#>  8:        8        0      0 <ResampleResult>
#>  9:        9        0      0 <ResampleResult>
#> 10:       10        0      0 <ResampleResult>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))