Hyperparameter Tuning with Non-linear Optimization

Subclass for non-linear optimization (NLopt). Calls nloptr::nloptr from package nloptr.

Source

Johnson, G S (2020). “The NLopt nonlinear-optimization package.” https://github.com/stevengj/nlopt.

Details

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the bbotk::Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("nloptr")

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerNLoptr which can be applied on any black box optimization problem. See also the documentation of bbotk.

Parameters

algorithm

character(1)

eval_g_ineq

function()

xtol_rel

numeric(1)

xtol_abs

numeric(1)

ftol_rel

numeric(1)

ftol_abs

numeric(1)

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Resources

book section on tuners.

mlr3hyperband extension package for the Hyperband algorithm.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerNLoptr

Methods

Public methods

TunerNLoptr$new()

TunerNLoptr$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerNLoptr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerNLoptr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization
# \donttest{

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  method = tnr("nloptr", algorithm = "NLOPT_LN_BOBYQA"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce")
)

# best performing hyperparameter configuration
instance$result
#>           cp learner_param_vals  x_domain classif.ce
#> 1: -4.890684          <list[2]> <list[1]> 0.06956522

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp classif.ce x_domain_cp runtime_learners           timestamp
#>  1: -4.890684 0.06956522 0.007516282            0.011 2022-11-17 17:28:07
#>  2: -4.890684 0.06956522 0.007516282            0.012 2022-11-17 17:28:07
#>  3: -4.890684 0.06956522 0.007516282            0.011 2022-11-17 17:28:08
#>  4: -3.163745 0.06956522 0.042267159            0.009 2022-11-17 17:28:08
#>  5: -6.617623 0.06956522 0.001336605            0.011 2022-11-17 17:28:08
#>  6: -4.873414 0.06956522 0.007647211            0.015 2022-11-17 17:28:08
#>  7: -4.907953 0.06956522 0.007387595            0.012 2022-11-17 17:28:08
#>  8: -4.888957 0.06956522 0.007529273            0.012 2022-11-17 17:28:08
#>  9: -4.892411 0.06956522 0.007503313            0.011 2022-11-17 17:28:08
#> 10: -4.890684 0.06956522 0.007516282            0.011 2022-11-17 17:28:08
#>     batch_nr warnings errors      resample_result
#>  1:        1        0      0 <ResampleResult[21]>
#>  2:        2        0      0 <ResampleResult[21]>
#>  3:        3        0      0 <ResampleResult[21]>
#>  4:        4        0      0 <ResampleResult[21]>
#>  5:        5        0      0 <ResampleResult[21]>
#>  6:        6        0      0 <ResampleResult[21]>
#>  7:        7        0      0 <ResampleResult[21]>
#>  8:        8        0      0 <ResampleResult[21]>
#>  9:        9        0      0 <ResampleResult[21]>
#> 10:       10        0      0 <ResampleResult[21]>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))
# }