Hyperparameter Tuning with Random Search

Subclass for random search tuning.

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281–305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Details

The random points are sampled by paradox::generate_design_random().

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("random_search")

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria. A batch contains of batch_size times resampling$iters jobs. E.g., if you set a batch size of 10 points and do a 5-fold cross validation, you can utilize up to 50 cores.

Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerBatchRandomSearch which can be applied on any black box optimization problem. See also the documentation of bbotk.

Parameters

batch_size

integer(1)
Maximum number of points to try in a batch.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Getting started with hyperparameter optimization.

An overview of all tuners can be found on our website.

Tune a support vector machine on the Sonar data set.

Learn about tuning spaces.

Estimate the model performance with nested resampling.

Learn about multi-objective optimization.

Simultaneously optimize hyperparameters and use early stopping with XGBoost.

Automate the tuning.

The gallery features a collection of case studies and demos about optimization.

Learn more advanced methods with the Practical Tuning Series.

Learn about hotstarting models.

Run the default hyperparameter configuration of learners as a baseline.

Use the Hyperband optimizer with different budget parameters.

The cheatsheet summarizes the most important functions of mlr3tuning.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerBatch -> mlr3tuning::TunerBatchFromOptimizerBatch -> TunerBatchRandomSearch

Methods

Public methods

TunerBatchRandomSearch$new()

TunerBatchRandomSearch$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerBatchFromOptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerBatchRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerBatchRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  tuner = tnr("random_search"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10
)

# best performing hyperparameter configuration
instance$result
#>           cp learner_param_vals  x_domain classif.ce
#>        <num>             <list>    <list>      <num>
#> 1: -8.159194          <list[2]> <list[1]> 0.04347826

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp classif.ce runtime_learners           timestamp warnings errors
#>         <num>      <num>            <num>              <POSc>    <int>  <int>
#>  1: -3.980719 0.05217391            0.005 2025-09-16 13:33:09        0      0
#>  2: -8.159194 0.04347826            0.023 2025-09-16 13:33:09        0      0
#>  3: -3.406518 0.05217391            0.006 2025-09-16 13:33:09        0      0
#>  4: -4.868614 0.04347826            0.006 2025-09-16 13:33:09        0      0
#>  5: -5.262721 0.04347826            0.006 2025-09-16 13:33:09        0      0
#>  6: -8.627951 0.04347826            0.005 2025-09-16 13:33:09        0      0
#>  7: -4.833278 0.05217391            0.006 2025-09-16 13:33:09        0      0
#>  8: -2.444900 0.05217391            0.006 2025-09-16 13:33:09        0      0
#>  9: -2.513168 0.05217391            0.006 2025-09-16 13:33:09        0      0
#> 10: -5.243984 0.04347826            0.007 2025-09-16 13:33:09        0      0
#>      x_domain batch_nr  resample_result
#>        <list>    <int>           <list>
#>  1: <list[1]>        1 <ResampleResult>
#>  2: <list[1]>        2 <ResampleResult>
#>  3: <list[1]>        3 <ResampleResult>
#>  4: <list[1]>        4 <ResampleResult>
#>  5: <list[1]>        5 <ResampleResult>
#>  6: <list[1]>        6 <ResampleResult>
#>  7: <list[1]>        7 <ResampleResult>
#>  8: <list[1]>        8 <ResampleResult>
#>  9: <list[1]>        9 <ResampleResult>
#> 10: <list[1]>       10 <ResampleResult>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))