Syntactic Sugar for Tuning Instance Construction

Function to construct a TuningInstanceSingleCrit or TuningInstanceMultiCrit.

Usage

ti(
  task,
  learner,
  resampling,
  measures = NULL,
  terminator,
  search_space = NULL,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE,
  allow_hotstart = FALSE,
  keep_hotstart_stack = FALSE,
  evaluate_default = FALSE,
  callbacks = list()
)

Arguments

task

(mlr3::Task)
Task to operate on.

learner

(mlr3::Learner)
Learner to tune.

resampling

(mlr3::Resampling)
Resampling that is used to evaluate the performance of the hyperparameter configurations. Uninstantiated resamplings are instantiated during construction so that all configurations are evaluated on the same data splits. Already instantiated resamplings are kept unchanged. Specialized Tuner change the resampling e.g. to evaluate a hyperparameter configuration on different data splits. This field, however, always returns the resampling passed in construction.

measures

(mlr3::Measure or list of mlr3::Measure)
A single measure creates a TuningInstanceSingleCrit and multiple measures a TuningInstanceMultiCrit. If NULL, default measure is used.

terminator

(Terminator)
Stop criterion of the tuning process.

search_space

(paradox::ParamSet)
Hyperparameter search space. If NULL (default), the search space is constructed from the TuneToken of the learner's parameter set (learner$param_set).

store_benchmark_result

(logical(1))
If TRUE (default), store resample result of evaluated hyperparameter configurations in archive as mlr3::BenchmarkResult.

store_models

(logical(1))
If TRUE, fitted models are stored in the benchmark result (archive$benchmark_result). If store_benchmark_result = FALSE, models are only stored temporarily and not accessible after the tuning. This combination is needed for measures that require a model.

check_values

(logical(1))
If TRUE, hyperparameter values are checked before evaluation and performance scores after. If FALSE (default), values are unchecked but computational overhead is reduced.

allow_hotstart

(logical(1))
Allow to hotstart learners with previously fitted models. See also mlr3::HotstartStack. The learner must support hotstarting. Sets store_models = TRUE.

keep_hotstart_stack

(logical(1))
If TRUE, mlr3::HotstartStack is kept in $objective$hotstart_stack after tuning.

evaluate_default

(logical(1))
If TRUE, learner is evaluated with hyperparameters set to their default values at the start of the optimization.

callbacks

(list of CallbackTuning)
List of callbacks.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Getting started with hyperparameter optimization.

Tune a simple classification tree on the Palmer Penguins data set.

Learn about tuning spaces.

The gallery features a collection of case studies and demos about optimization.

Learn more advanced methods with the practical tuning series.

Simultaneously optimize hyperparameters and use early stopping with XGBoost.

Make us of proven search space.

Learn about hotstarting models.

Run the default hyperparameter configuration of learners as a baseline.

Default Measures

If no measure is passed, the default measure is used. The default measure depends on the task type.

Task	Default Measure	Package

"classif"	"classif.ce"	mlr3

"regr"	"regr.mse"	mlr3

"surv"	"surv.cindex"	mlr3proba

"dens"	"dens.logloss"	mlr3proba

"classif_st"	"classif.ce"	mlr3spatial

"regr_st"	"regr.mse"	mlr3spatial

"clust"	"clust.dunn"	mlr3cluster

Examples

# Hyperparameter optimization on the Palmer Penguins data set
task = tsk("penguins")

# Load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# Construct tuning instance
instance = ti(
  task = task,
  learner = learner,
  resampling = rsmp("cv", folds = 3),
  measures = msr("classif.ce"),
  terminator = trm("evals", n_evals = 4)
)

# Choose optimization algorithm
tuner = tnr("random_search", batch_size = 2)

# Run tuning
tuner$optimize(instance)
#>           cp learner_param_vals  x_domain classif.ce
#> 1: -3.163854          <list[2]> <list[1]> 0.06399695

# Set optimal hyperparameter configuration to learner
learner$param_set$values = instance$result_learner_param_vals

# Train the learner on the full data set
learner$train(task)

# Inspect all evaluated configurations
as.data.table(instance$archive)
#>           cp classif.ce x_domain_cp runtime_learners           timestamp
#> 1: -3.163854 0.06399695 0.042262562            0.024 2023-03-09 07:19:12
#> 2: -6.432404 0.06399695 0.001608578            0.022 2023-03-09 07:19:12
#> 3: -6.808200 0.06399695 0.001104680            0.022 2023-03-09 07:19:12
#> 4: -2.664051 0.06399695 0.069665438            0.021 2023-03-09 07:19:12
#>    batch_nr warnings errors      resample_result
#> 1:        1        0      0 <ResampleResult[21]>
#> 2:        1        0      0 <ResampleResult[21]>
#> 3:        2        0      0 <ResampleResult[21]>
#> 4:        2        0      0 <ResampleResult[21]>