AutoTuner

The AutoTuner is a mlr3::Learner which wraps another mlr3::Learner and performs the following steps during $train():

The hyperparameters of the wrapped (inner) learner are trained on the training data via resampling. The tuning can be specified by providing a Tuner, a bbotk::Terminator, a search space as paradox::ParamSet, a mlr3::Resampling and a mlr3::Measure.

The best found hyperparameter configuration is set as hyperparameters for the wrapped (inner) learner.

A final model is fit on the complete training data using the now parametrized wrapped learner.

During $predict() the AutoTuner just calls the predict method of the wrapped (inner) learner.

Note that this approach allows to perform nested resampling by passing an AutoTuner object to mlr3::resample() or mlr3::benchmark(). To access the inner resampling results, set store_tuning_instance = TRUE and execute mlr3::resample() or mlr3::benchmark() with store_models = TRUE (see examples).

Super class

mlr3::Learner -> AutoTuner

Public fields

instance_args

(list())
All arguments from construction to create the TuningInstanceSingleCrit.

tuner

(Tuner).

Active bindings

archive

ArchiveTuning
Archive of the TuningInstanceSingleCrit.

learner

(mlr3::Learner)
Trained learner

tuning_instance

(TuningInstanceSingleCrit)
Internally created tuning instance with all intermediate results.

tuning_result

(named list())
Short-cut to result from TuningInstanceSingleCrit.

param_set

paradox::ParamSet.

predict_type

(character(1))
Stores the currently active predict type, e.g. "response". Must be an element of $predict_types.

Methods

Public methods

AutoTuner$new()

AutoTuner$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

AutoTuner$new(
  learner,
  resampling,
  measure,
  search_space,
  terminator,
  tuner,
  store_tuning_instance = TRUE,
  store_benchmark_result = TRUE,
  store_models = FALSE,
  check_values = FALSE
)

Arguments

learner

(mlr3::Learner)
Learner to tune, see TuningInstanceSingleCrit.

resampling

(mlr3::Resampling)
Resampling strategy during tuning, see TuningInstanceSingleCrit. This mlr3::Resampling is meant to be the inner resampling, operating on the training set of an arbitrary outer resampling. For this reason it is not feasible to pass an instantiated mlr3::Resampling here.

measure

(mlr3::Measure)
Performance measure to optimize.

search_space

(paradox::ParamSet)
Hyperparameter search space, see TuningInstanceSingleCrit.

terminator

(bbotk::Terminator)
When to stop tuning, see TuningInstanceSingleCrit.

tuner

(Tuner)
Tuning algorithm to run.

store_tuning_instance

(logical(1))
If TRUE (default), stores the internally created TuningInstanceSingleCrit with all intermediate results in slot $tuning_instance.

store_benchmark_result

(logical(1))
Store benchmark result in archive?

store_models

(logical(1))
Store models in benchmark result?

check_values

(logical(1))
Should parameters before the evaluation and the results be checked for validity?

Method clone()

The objects of this class are cloneable with this method.

Usage

AutoTuner$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(paradox)
task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measure = msr("classif.ce")
search_space = ParamSet$new(
  params = list(ParamDbl$new("cp", lower = 0.001, upper = 0.1)))

terminator = trm("evals", n_evals = 5)
tuner = tnr("grid_search")
at = AutoTuner$new(
  learner, resampling, measure, search_space, terminator,
  tuner, store_tuning_instance = TRUE)

at$train(task)
at$model

#> $learner
#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0, cp=0.056
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
#> $tuning_instance
#> <TuningInstanceSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveTuning:classif.rpart_on_iris>
#> * Search Space:
#> <ParamSet>
#>    id    class lower upper levels        default value
#> 1: cp ParamDbl 0.001   0.1        <NoDefault[3]>      
#> * Terminator: <TerminatorEvals>
#> * Terminated: TRUE
#> * Result:
#>       cp learner_param_vals  x_domain classif.ce
#> 1: 0.056          <list[2]> <list[1]>       0.06
#> * Archive:
#> <ArchiveTuning>
#>       cp classif.ce                                uhash  x_domain
#> 1: 0.056       0.06 f0567a7f-edb1-4540-8a9d-db6aa4005456 <list[1]>
#> 2: 0.045       0.06 54e8b592-05cf-4308-9e53-43c3edb80ab1 <list[1]>
#> 3: 0.023       0.06 fa5f4dec-cb38-4614-84ca-e9fbb00d115a <list[1]>
#> 4: 0.012       0.06 1d12d807-4687-4fe1-b520-470077ed9e49 <list[1]>
#> 5: 0.089       0.06 fbf07079-230b-4fe6-b4bd-f7391576b4ff <list[1]>
#>              timestamp batch_nr
#> 1: 2020-10-28 10:02:14        1
#> 2: 2020-10-28 10:02:14        2
#> 3: 2020-10-28 10:02:14        3
#> 4: 2020-10-28 10:02:14        4
#> 5: 2020-10-28 10:02:15        5
#> 

at$learner

#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0, cp=0.056
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights

# Nested resampling
at = AutoTuner$new(learner, resampling, measure, search_space, terminator,
  tuner, store_tuning_instance = TRUE)

resampling_outer = rsmp("cv", folds = 2)
rr = resample(task, at, resampling_outer, store_models = TRUE)

# Aggregate performance of outer results
rr$aggregate()

#> classif.ce 
#> 0.09333333 

# Retrieve inner tuning results.
as.data.table(rr)$learner[[1]]$tuning_result

#>       cp learner_param_vals  x_domain classif.ce
#> 1: 0.056          <list[2]> <list[1]>       0.04