AutoTuner

The AutoTuner is a mlr3::Learner which wraps another mlr3::Learner and performs the following steps during $train():

The hyperparameters of the wrapped (inner) learner are trained on the training data via resampling. The tuning can be specified by providing a Tuner, a bbotk::Terminator, a search space as paradox::ParamSet, a mlr3::Resampling and a mlr3::Measure.

The best found hyperparameter configuration is set as hyperparameters for the wrapped (inner) learner.

A final model is fit on the complete training data using the now parametrized wrapped learner.

During $predict() the AutoTuner just calls the predict method of the wrapped (inner) learner.

Note that this approach allows to perform nested resampling by passing an AutoTuner object to mlr3::resample() or mlr3::benchmark(). To access the inner resampling results, set store_tuning_instance = TRUE and execute mlr3::resample() or mlr3::benchmark() with store_models = TRUE (see examples).

Super class

mlr3::Learner -> AutoTuner

Public fields

instance_args

(list())
All arguments from construction to create the TuningInstanceSingleCrit.

tuner

(Tuner).

store_tuning_instance

(logical(1))
If TRUE (default), stores the internally created TuningInstanceSingleCrit with all intermediate results in slot $tuning_instance.

Active bindings

archive

ArchiveTuning
Archive of the TuningInstanceSingleCrit.

learner

(mlr3::Learner)
Trained learner

tuning_instance

(TuningInstanceSingleCrit)
Internally created tuning instance with all intermediate results.

tuning_result

(named list())
Short-cut to result from TuningInstanceSingleCrit.

param_set

paradox::ParamSet.

Methods

Public methods

AutoTuner$new()

AutoTuner$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

AutoTuner$new(learner, resampling, measure, search_space, terminator, tuner)

Arguments

learner

(mlr3::Learner)
Learner to tune, see TuningInstanceSingleCrit.

resampling

(mlr3::Resampling)
Resampling strategy during tuning, see TuningInstanceSingleCrit. This mlr3::Resampling is meant to be the inner resampling, operating on the training set of an arbitrary outer resampling. For this reason it is not feasible to pass an instantiated mlr3::Resampling here.

measure

(list of mlr3::Measure)
Performance measure to optimize.

search_space

(paradox::ParamSet)
Hyperparameter search space, see TuningInstanceSingleCrit.

terminator

(bbotk::Terminator)
When to stop tuning, see TuningInstanceSingleCrit.

tuner

(Tuner)
Tuning algorithm to run.

Method clone()

The objects of this class are cloneable with this method.

Usage

AutoTuner$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(paradox)
task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measure = msr("classif.ce")
search_space = ParamSet$new(
  params = list(ParamDbl$new("cp", lower = 0.001, upper = 0.1)))

terminator = trm("evals", n_evals = 5)
tuner = tnr("grid_search")
at = AutoTuner$new(
  learner, resampling, measure, search_space, terminator,
  tuner)
at$store_tuning_instance = TRUE

at$train(task)
at$model

#> $learner
#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0, cp=0.056
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
#> $tuning_instance
#> <TuningInstanceSingleCrit>
#> * State:  Optimized
#> * Objective: <ObjectiveTuning:classif.rpart_on_iris>
#> * Search Space:
#> <ParamSet>
#>    id    class lower upper levels        default value
#> 1: cp ParamDbl 0.001   0.1        <NoDefault[3]>      
#> * Terminator: <TerminatorEvals>
#> * Terminated: TRUE
#> * Result:
#>       cp learner_param_vals  x_domain classif.ce
#> 1: 0.056          <list[2]> <list[1]>       0.06
#> * Archive:
#> <ArchiveTuning>
#>       cp classif.ce                                uhash  x_domain
#> 1: 0.056       0.06 39e74d5c-1cd9-4110-8867-9568a9a49159 <list[1]>
#> 2: 0.045       0.06 a42002f3-8667-4baf-9527-f6a9624190cb <list[1]>
#> 3: 0.023       0.06 5d55b611-bd1d-414e-a533-87d9f6d0fc1d <list[1]>
#> 4: 0.012       0.06 c79af703-84e2-49d9-bf90-2bad84262064 <list[1]>
#> 5: 0.089       0.06 0ed326fe-628c-4579-8bd0-88abb2716e52 <list[1]>
#>              timestamp batch_nr
#> 1: 2020-08-16 04:26:32        1
#> 2: 2020-08-16 04:26:33        2
#> 3: 2020-08-16 04:26:33        3
#> 4: 2020-08-16 04:26:33        4
#> 5: 2020-08-16 04:26:33        5
#> 

at$learner

#> <LearnerClassifRpart:classif.rpart>
#> * Model: rpart
#> * Parameters: xval=0, cp=0.056
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights

# Nested resampling
at = AutoTuner$new(learner, resampling, measure, search_space, terminator,
  tuner)
at$store_tuning_instance = TRUE

resampling_outer = rsmp("cv", folds = 2)
rr = resample(task, at, resampling_outer, store_models = TRUE)

# Aggregate performance of outer results
rr$aggregate()

#> classif.ce 
#> 0.05333333 

# Retrieve inner tuning results.
rr$data$learner[[1]]$tuning_result

#>       cp learner_param_vals  x_domain classif.ce
#> 1: 0.056          <list[2]> <list[1]>       0.08