Tuner

The Tuner implements the optimization algorithm.

Details

Tuner is an abstract base class that implements the base functionality each tuner must provide.

Extension Packages

Additional tuners are provided by the following packages.

mlr3hyperband adds the Hyperband and Successive Halving algorithm.

mlr3mbo adds Bayesian optimization methods.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Getting started with hyperparameter optimization.

An overview of all tuners can be found on our website.

Tune a support vector machine on the Sonar data set.

Learn about tuning spaces.

Estimate the model performance with nested resampling.

Learn about multi-objective optimization.

Simultaneously optimize hyperparameters and use early stopping with XGBoost.

Automate the tuning.

The gallery features a collection of case studies and demos about optimization.

Learn more advanced methods with the Practical Tuning Series.

Learn about hotstarting models.

Run the default hyperparameter configuration of learners as a baseline.

Use the Hyperband optimizer with different budget parameters.

The cheatsheet summarizes the most important functions of mlr3tuning.

See also

Public fields

id

(character(1))
Identifier of the object. Used in tables, plot and text output.

Active bindings

param_set

(paradox::ParamSet)
Set of control parameters.

param_classes

(character())
Supported parameter classes for learner hyperparameters that the tuner can optimize, as given in the paradox::ParamSet $class field.

properties

(character())
Set of properties of the tuner. Must be a subset of mlr_reflections$tuner_properties.

packages

(character())
Set of required packages. Note that these packages will be loaded via requireNamespace(), and are not attached.

label

(character(1))
Label for this object. Can be used in tables, plot and text output instead of the ID.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Methods

Public methods

Tuner$new()

Tuner$format()

Tuner$print()

Tuner$help()

Tuner$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Tuner$new(
  id = "tuner",
  param_set,
  param_classes,
  properties,
  packages = character(),
  label = NA_character_,
  man = NA_character_
)

Arguments

id

(character(1))
Identifier for the new instance.

param_set

(paradox::ParamSet)
Set of control parameters.

param_classes

(character())
Supported parameter classes for learner hyperparameters that the tuner can optimize, as given in the paradox::ParamSet $class field.

properties

(character())
Set of properties of the tuner. Must be a subset of mlr_reflections$tuner_properties.

packages

(character())
Set of required packages. Note that these packages will be loaded via requireNamespace(), and are not attached.

label

(character(1))
Label for this object. Can be used in tables, plot and text output instead of the ID.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method format()

Helper for print outputs.

Usage

Tuner$format(...)

Arguments

...

(ignored).

Returns

(character()).

Method print()

Print method.

Usage

Tuner$print()

Returns

(character()).

Method help()

Opens the corresponding help page referenced by field $man.

Usage

Tuner$help()

Method clone()

The objects of this class are cloneable with this method.

Usage

Tuner$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.