TuningInstanceSingleCrit

Specifies a general tuning scenario, including objective function and archive for Tuners to act upon. This class stores an ObjectiveTuning object that encodes the black box objective function which a Tuner has to optimize. It allows the basic operations of querying the objective at design points ($eval_batch()), storing the evaluations in the internal Archive and accessing the final result ($result).

Evaluations of hyperparameter configurations are performed in batches by calling mlr3::benchmark() internally. Before a batch is evaluated, the Terminator is queried for the remaining budget. If the available budget is exhausted, an exception is raised, and no further evaluations can be performed from this point on.

The tuner is also supposed to store its final result, consisting of a selected hyperparameter configuration and associated estimated performance values, by calling the method instance$assign_result.

Super classes

bbotk::OptimInstance -> bbotk::OptimInstanceSingleCrit -> TuningInstanceSingleCrit

Active bindings

result_learner_param_vals

(list()) Param values for the optimal learner call.

Methods

Public methods

TuningInstanceSingleCrit$new()

TuningInstanceSingleCrit$assign_result()

TuningInstanceSingleCrit$clone()

Inherited methods

bbotk::OptimInstance$eval_batch()

bbotk::OptimInstance$format()

bbotk::OptimInstance$objective_function()

bbotk::OptimInstance$print()

Method new()

Creates a new instance of this R6 class.

This defines the resampled performance of a learner on a task, a feasibility region for the parameters the tuner is supposed to optimize, and a termination criterion.

Usage

TuningInstanceSingleCrit$new(
  task,
  learner,
  resampling,
  measure,
  search_space,
  terminator,
  store_models = FALSE
)

Arguments

task

(mlr3::Task)

learner

(mlr3::Learner)

resampling

(mlr3::Resampling)
Note that uninstantiated resamplings are instantiated during construction so that all configurations are evaluated on the same data splits.

measure

(mlr3::Measure)
Measure to optimize.

search_space

(paradox::ParamSet)

terminator

(Terminator)

store_models

logical(1)

Method assign_result()

The Tuner object writes the best found point and estimated performance value here. For internal use.

Usage

TuningInstanceSingleCrit$assign_result(xdt, y, learner_param_vals = NULL)

Arguments

xdt

(data.table)
x values as data.table with one row. Contains the value in the search space of the TuningInstanceSingleCrit object. Can contain additional columns for extra information.

y

(numeric(1))
Optimal outcome.

learner_param_vals

(list())
Fixed parameter values of the learner that are neither part of the

Method clone()

The objects of this class are cloneable with this method.

Usage

TuningInstanceSingleCrit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)
library(paradox)
library(mlr3)

# Objects required to define the performance evaluator:
task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measure = msr("classif.ce")
param_set = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1),
  ParamInt$new("minsplit", lower = 1, upper = 10))
)

terminator = term("evals", n_evals = 5)
inst = TuningInstanceSingleCrit$new(
  task = task,
  learner = learner,
  resampling = resampling,
  measure = measure,
  search_space = param_set,
  terminator = terminator
)

# first 4 points as cross product
design = CJ(cp = c(0.05, 0.01), minsplit = c(5, 3))
inst$eval_batch(design)
inst$archive

#> <Archive>
#>      cp minsplit classif.ce  resample_result x_domain           timestamp
#> 1: 0.01        3       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#> 2: 0.01        5       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#> 3: 0.05        3       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#> 4: 0.05        5       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#>    batch_nr
#> 1:        1
#> 2:        1
#> 3:        1
#> 4:        1

# try more points, catch the raised terminated message
tryCatch(
  inst$eval_batch(data.table(cp = 0.01, minsplit = 7)),
  terminated_error = function(e) message(as.character(e))
)

# try another point although the budget is now exhausted
# -> no extra evaluations
tryCatch(
  inst$eval_batch(data.table(cp = 0.01, minsplit = 9)),
  terminated_error = function(e) message(as.character(e))
)

#> Error: Objective (obj:classif.rpart_on_iris, term:<TerminatorEvals>) terminated

inst$archive

#> <Archive>
#>      cp minsplit classif.ce  resample_result x_domain           timestamp
#> 1: 0.01        3       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#> 2: 0.01        5       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#> 3: 0.05        3       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#> 4: 0.05        5       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#> 5: 0.01        7       0.04 <ResampleResult>   <list> 2020-07-05 04:26:27
#>    batch_nr
#> 1:        1
#> 2:        1
#> 3:        1
#> 4:        1
#> 5:        2

### Error handling
# get a learner which breaks with 50% probability
# set encapsulation + fallback
learner = lrn("classif.debug", error_train = 0.5)
learner$encapsulate = c(train = "evaluate", predict = "evaluate")
learner$fallback = lrn("classif.featureless")

param_set = ParamSet$new(list(
  ParamDbl$new("x", lower = 0, upper = 1)
))

inst = TuningInstanceSingleCrit$new(
  task = tsk("wine"),
  learner = learner,
  resampling = rsmp("cv", folds = 3),
  measure = msr("classif.ce"),
  search_space = param_set,
  terminator = term("evals", n_evals = 5)
)

tryCatch(
  inst$eval_batch(data.table(x = 1:5 / 5)),
  terminated_error = function(e) message(as.character(e))
)

archive = inst$archive$data()

# column errors: multiple errors recorded
print(archive)

#>      x classif.ce  resample_result x_domain           timestamp batch_nr
#> 1: 0.2  0.6401130 <ResampleResult>   <list> 2020-07-05 04:26:28        1
#> 2: 0.4  0.6068738 <ResampleResult>   <list> 2020-07-05 04:26:28        1
#> 3: 0.6  0.6456685 <ResampleResult>   <list> 2020-07-05 04:26:28        1
#> 4: 0.8  0.6068738 <ResampleResult>   <list> 2020-07-05 04:26:28        1
#> 5: 1.0  0.6238230 <ResampleResult>   <list> 2020-07-05 04:26:28        1