TunerCmaes

Subclass that implements CMA-ES calling adagio::pureCMAES() from package adagio.

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

mlr_tuners$get("cmaes")
tnr("cmaes")

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Parameters

par

numeric()

sigma

numeric(1)

For the meaning of the control parameters, see adagio::pureCMAES(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerCmaes

Methods

Public methods

TunerCmaes$new()

TunerCmaes$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerCmaes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerCmaes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(paradox)
library(data.table)
search_space = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1)
))
terminator = trm("evals", n_evals = 10)
instance = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  search_space = search_space,
  terminator = terminator
)
tt = tnr("cmaes", par = 0.1)
# modifies the instance by reference
tt$optimize(instance)

#>     cp learner_param_vals  x_domain classif.ce
#> 1: 0.1          <list[2]> <list[1]>        0.1

# returns best configuration and best performance
instance$result

#>     cp learner_param_vals  x_domain classif.ce
#> 1: 0.1          <list[2]> <list[1]>        0.1

# allows access of data.table of full path of all evaluations
instance$archive

#> <ArchiveTuning>
#>             cp classif.ce                                uhash  x_domain
#>  1: 0.10000000        0.1 cf36b3e6-fbbc-4736-8fc4-1b2030471864 <list[1]>
#>  2: 0.10000000        0.1 11466459-c61c-4efb-850d-2b1a56abb8f9 <list[1]>
#>  3: 0.10000000        0.1 396c8e06-beb7-4b88-bbf0-ee66812162f3 <list[1]>
#>  4: 0.10000000        0.1 7f50faf4-51a7-422b-b988-7cc5a5a29ced <list[1]>
#>  5: 0.10000000        0.1 d6d561a4-1124-449e-b17b-cd6cbdb0ee2f <list[1]>
#>  6: 0.09357766        0.1 f37358b7-7786-4298-acff-81f577f1fbac <list[1]>
#>  7: 0.08681602        0.1 92375380-a124-458f-b73c-a70daa8d4d44 <list[1]>
#>  8: 0.10000000        0.1 b4a0f476-f2c9-480c-9139-69bbdcbcdbea <list[1]>
#>  9: 0.08461986        0.1 34aa7342-6101-4005-b4d8-9e18838b4c38 <list[1]>
#> 10: 0.10000000        0.1 fb89ca17-f553-43ef-9e7d-480f50258462 <list[1]>
#>               timestamp batch_nr
#>  1: 2020-12-12 10:07:51        1
#>  2: 2020-12-12 10:07:52        2
#>  3: 2020-12-12 10:07:52        3
#>  4: 2020-12-12 10:07:52        4
#>  5: 2020-12-12 10:07:52        5
#>  6: 2020-12-12 10:07:52        6
#>  7: 2020-12-12 10:07:52        7
#>  8: 2020-12-12 10:07:52        8
#>  9: 2020-12-12 10:07:52        9
#> 10: 2020-12-12 10:07:53       10