Hyperparameter Tuning with Covariance Matrix Adaptation Evolution Strategy

Subclass for Covariance Matrix Adaptation Evolution Strategy (CMA-ES). Calls adagio::pureCMAES() from package adagio.

Source

Hansen N (2016). “The CMA Evolution Strategy: A Tutorial.” 1604.00772.

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("cmaes")

Control Parameters

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see adagio::pureCMAES(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerBatchCmaes which can be applied on any black box optimization problem. See also the documentation of bbotk.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

An overview of all tuners can be found on our website.

Learn more about tuners.

The gallery features a collection of case studies and demos about optimization.

Use the Hyperband optimizer with different budget parameters.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerBatch -> mlr3tuning::TunerBatchFromOptimizerBatch -> TunerBatchCmaes

Methods

Public methods

TunerBatchCmaes$new()

TunerBatchCmaes$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerBatchFromOptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerBatchCmaes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerBatchCmaes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE),
  minsplit = to_tune(p_dbl(2, 128, trafo = as.integer)),
  minbucket = to_tune(p_dbl(1, 64, trafo = as.integer))
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  tuner = tnr("cmaes"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10)

# best performing hyperparameter configuration
instance$result
#>           cp minbucket minsplit learner_param_vals  x_domain classif.ce
#>        <num>     <num>    <num>             <list>    <list>      <num>
#> 1: -6.284677  53.88483 68.59505          <list[4]> <list[3]> 0.03478261

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp minbucket  minsplit classif.ce  x_domain_cp x_domain_minbucket
#>         <num>     <num>     <num>      <num>        <num>              <int>
#>  1: -7.551355  64.00000  54.48565 0.11304348 0.0005253975                 64
#>  2: -6.284677  53.88483  68.59505 0.03478261 0.0018646590                 53
#>  3: -5.425148  54.39101   2.00000 0.04347826 0.0044044149                 54
#>  4: -2.302585  47.96519  85.40966 0.04347826 0.1000000000                 47
#>  5: -5.076769  64.00000  18.85283 0.11304348 0.0062400377                 64
#>  6: -6.192543  49.38410 128.00000 0.03478261 0.0020446203                 49
#>  7: -6.466042  64.00000  74.86460 0.11304348 0.0015553691                 64
#>  8: -6.184380  47.63568   2.00000 0.04347826 0.0020613792                 47
#>  9: -2.302585  30.22800 128.00000 0.04347826 0.1000000000                 30
#> 10: -3.632083  59.33941  76.25020 0.06956522 0.0264610179                 59
#>     x_domain_minsplit runtime_learners           timestamp batch_nr warnings
#>                 <int>            <num>              <POSc>    <int>    <int>
#>  1:                54            0.009 2024-05-14 10:21:11        1        0
#>  2:                68            0.026 2024-05-14 10:21:12        2        0
#>  3:                 2            0.010 2024-05-14 10:21:12        3        0
#>  4:                85            0.010 2024-05-14 10:21:12        4        0
#>  5:                18            0.010 2024-05-14 10:21:12        5        0
#>  6:               128            0.010 2024-05-14 10:21:12        6        0
#>  7:                74            0.009 2024-05-14 10:21:12        7        0
#>  8:                 2            0.009 2024-05-14 10:21:12        8        0
#>  9:               128            0.011 2024-05-14 10:21:12        9        0
#> 10:                76            0.010 2024-05-14 10:21:12       10        0
#>     errors  resample_result
#>      <int>           <list>
#>  1:      0 <ResampleResult>
#>  2:      0 <ResampleResult>
#>  3:      0 <ResampleResult>
#>  4:      0 <ResampleResult>
#>  5:      0 <ResampleResult>
#>  6:      0 <ResampleResult>
#>  7:      0 <ResampleResult>
#>  8:      0 <ResampleResult>
#>  9:      0 <ResampleResult>
#> 10:      0 <ResampleResult>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))