Hyperparameter Tuning with Covariance Matrix Adaptation Evolution Strategy

Subclass for Covariance Matrix Adaptation Evolution Strategy (CMA-ES). Calls adagio::pureCMAES() from package adagio.

Source

Hansen N (2016). “The CMA Evolution Strategy: A Tutorial.” 1604.00772.

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("cmaes")

Control Parameters

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see adagio::pureCMAES(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerCmaes which can be applied on any black box optimization problem. See also the documentation of bbotk.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Learn more about tuners.

The gallery features a collection of case studies and demos about optimization.

Use the Hyperband optimizer with different budget parameters.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerCmaes

Methods

Public methods

TunerCmaes$new()

TunerCmaes$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerCmaes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerCmaes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE),
  minsplit = to_tune(p_dbl(2, 128, trafo = as.integer)),
  minbucket = to_tune(p_dbl(1, 64, trafo = as.integer))
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  method = tnr("cmaes"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 10)

# best performing hyperparameter configuration
instance$result
#>           cp minsplit minbucket learner_param_vals  x_domain classif.ce
#> 1: -2.302585      128  9.550805          <list[4]> <list[3]> 0.06956522

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp  minsplit minbucket classif.ce  x_domain_cp x_domain_minsplit
#>  1: -5.206836 120.34133 59.511262 0.07826087 0.0054789806               120
#>  2: -9.210340 128.00000 55.734128 0.07826087 0.0001000000               128
#>  3: -8.474953  63.97801 64.000000 0.12173913 0.0002086291                63
#>  4: -2.302585 128.00000  9.550805 0.06956522 0.1000000000               128
#>  5: -7.070500 121.81781  1.000000 0.06956522 0.0008498078               121
#>  6: -7.645283 123.19217 64.000000 0.12173913 0.0004782949               123
#>  7: -9.163435 128.00000 64.000000 0.12173913 0.0001048023               128
#>  8: -2.302585 128.00000 23.316485 0.06956522 0.1000000000               128
#>  9: -3.910971 128.00000 25.256877 0.06956522 0.0200210548               128
#> 10: -4.765959 128.00000 10.485188 0.06956522 0.0085147160               128
#>     x_domain_minbucket runtime_learners           timestamp batch_nr warnings
#>  1:                 59            0.009 2023-02-28 11:50:58        1        0
#>  2:                 55            0.009 2023-02-28 11:50:58        2        0
#>  3:                 64            0.008 2023-02-28 11:50:58        3        0
#>  4:                  9            0.007 2023-02-28 11:50:58        4        0
#>  5:                  1            0.009 2023-02-28 11:50:58        5        0
#>  6:                 64            0.010 2023-02-28 11:50:58        6        0
#>  7:                 64            0.009 2023-02-28 11:50:58        7        0
#>  8:                 23            0.008 2023-02-28 11:50:59        8        0
#>  9:                 25            0.007 2023-02-28 11:50:59        9        0
#> 10:                 10            0.008 2023-02-28 11:50:59       10        0
#>     errors      resample_result
#>  1:      0 <ResampleResult[21]>
#>  2:      0 <ResampleResult[21]>
#>  3:      0 <ResampleResult[21]>
#>  4:      0 <ResampleResult[21]>
#>  5:      0 <ResampleResult[21]>
#>  6:      0 <ResampleResult[21]>
#>  7:      0 <ResampleResult[21]>
#>  8:      0 <ResampleResult[21]>
#>  9:      0 <ResampleResult[21]>
#> 10:      0 <ResampleResult[21]>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))