TunerGenSA

Subclass for generalized simulated annealing tuning calling GenSA::GenSA() from package GenSA.

Format

R6::R6Class object inheriting from Tuner.

Construction

TunerGenSA$new()
tnr("gensa")

Parameters

smooth :: logical(1)

temperature :: numeric(1)

acceptance.param :: numeric(1)

verbose :: logical(1)

trace.mat :: logical(1)

For the meaning of the control parameters, see GenSA::GenSA(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

See also

Examples

# see ?Tuner