TunerGenSA

Subclass for generalized simulated annealing tuning calling GenSA::GenSA() from package GenSA.

Source

Tsallis C, Stariolo DA (1996). “Generalized simulated annealing.” Physica A: Statistical Mechanics and its Applications, 233(1-2), 395--406. doi: 10.1016/s0378-4371(96)00271-3 . Xiang Y, Gubian S, Suomela B, Hoeng J (2013). “Generalized Simulated Annealing for Global Optimization: The GenSA Package.” The R Journal, 5(1), 13. doi: 10.32614/rj-2013-002 .

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

mlr_tuners$get("gensa")
tnr("gensa")

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Parameters

smooth

logical(1)

temperature

numeric(1)

acceptance.param

numeric(1)

verbose

logical(1)

trace.mat

logical(1)

For the meaning of the control parameters, see GenSA::GenSA(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerGenSA

Methods

Public methods

TunerGenSA$new()

TunerGenSA$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerGenSA$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerGenSA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(paradox)
search_space = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1)
))
terminator = trm("evals", n_evals = 3)
instance = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  search_space = search_space,
  terminator = terminator
)
tt = tnr("gensa")

# modifies the instance by reference
tt$optimize(instance)

#>             cp learner_param_vals  x_domain classif.ce
#> 1: 0.003964699          <list[2]> <list[1]>       0.08

# returns best configuration and best performance
instance$result

#>             cp learner_param_vals  x_domain classif.ce
#> 1: 0.003964699          <list[2]> <list[1]>       0.08

# allows access of data.table of full path of all evaluations
instance$archive

#> <ArchiveTuning>
#>             cp classif.ce                                uhash  x_domain
#> 1: 0.003964699       0.08 0ff124a6-878b-486b-b77c-8e09e309a6ff <list[1]>
#> 2: 0.067427466       0.08 7686b303-3cc4-4594-ae44-fe2d8c69a432 <list[1]>
#> 3: 0.082436805       0.08 830c2e3d-3db0-4fed-8700-4eebf6662f47 <list[1]>
#>              timestamp batch_nr
#> 1: 2020-10-12 04:35:00        1
#> 2: 2020-10-12 04:35:00        2
#> 3: 2020-10-12 04:35:00        3