Hyperparameter Tuning with Generalized Simulated Annealing

Subclass for generalized simulated annealing tuning calling GenSA::GenSA() from package GenSA.

Source

Tsallis C, Stariolo DA (1996). “Generalized simulated annealing.” Physica A: Statistical Mechanics and its Applications, 233(1-2), 395--406. doi: 10.1016/s0378-4371(96)00271-3 .

Xiang Y, Gubian S, Suomela B, Hoeng J (2013). “Generalized Simulated Annealing for Global Optimization: The GenSA Package.” The R Journal, 5(1), 13. doi: 10.32614/rj-2013-002 .

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

TunerGenSA$new()
mlr_tuners$get("gensa")
tnr("gensa")

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Parameters

smooth

logical(1)

temperature

numeric(1)

acceptance.param

numeric(1)

verbose

logical(1)

trace.mat

logical(1)

For the meaning of the control parameters, see GenSA::GenSA(). Note that we have removed all control parameters which refer to the termination of the algorithm and where our terminators allow to obtain the same behavior.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerGenSA

Methods

Public methods

TunerGenSA$new()

TunerGenSA$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerGenSA$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerGenSA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(paradox)
search_space = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1)
))
terminator = trm("evals", n_evals = 3)
instance = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  search_space = search_space,
  terminator = terminator
)
tt = tnr("gensa")

# modifies the instance by reference
tt$optimize(instance)

#>            cp learner_param_vals  x_domain classif.ce
#> 1: 0.03673619          <list[2]> <list[1]>       0.02

# returns best configuration and best performance
instance$result

#>            cp learner_param_vals  x_domain classif.ce
#> 1: 0.03673619          <list[2]> <list[1]>       0.02

# allows access of data.table of full path of all evaluations
instance$archive

#> <ArchiveTuning>
#>        cp classif.ce           timestamp batch_nr
#> 1: 0.0367       0.02 2021-04-09 04:26:35        1
#> 2: 0.0012       0.02 2021-04-09 04:26:35        2
#> 3: 0.0162       0.02 2021-04-09 04:26:35        3