Tuning via Iterated Racing.

TunerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi: 10.1016/j.orp.2016.09.002 .

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

TunerIrace$new()
mlr_tuners$get("irace")
tnr("irace")

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Parameters

n_instances

integer(1)
Number of resampling instances.

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorRunTime or TerminatorEvals instead. Other terminators do not work with TunerIrace. We substract 5 seconds from the TerminatorRunTime budget for stability reasons.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

See also

Super class

mlr3tuning::Tuner -> TunerIrace

Methods

Public methods

TunerIrace$new()

TunerIrace$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$optimize()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner and set search space
learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE))

# hyperparameter tuning on the pima indians diabetes data set
instance = tune(
  method = "irace",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 42
)

#> # 2021-07-13 05:09:28 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 2
#> # minNbSurvival: 2
#> # nbParameters: 1
#> # seed: 983633176
#> # confidence level: 0.95
#> # budget: 42
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2021-07-13 05:09:28 UTC: Iteration 1 of 2
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 42
#> # currentBudget: 21
#> # nbConfigurations: 3
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|          3|          3|   0.2578125000|          3|00:00:00|   NA|  NA|    NA|
#> |x|          2|          3|          3|   0.2656250000|          6|00:00:00|+0.86|0.93|0.1627|
#> |x|          3|          3|          3|   0.2513020833|          9|00:00:00|+0.90|0.93|0.1085|
#> |x|          4|          3|          3|   0.2529296875|         12|00:00:00|+0.90|0.93|0.1165|
#> |-|          5|          1|          3|   0.2437500000|         15|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           3    mean value:    0.2437500000
#> Description of the best-so-far configuration:
#>   .ID.     cp .PARENT.
#> 3    3 -3.696       NA
#> 
#> # 2021-07-13 05:09:30 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>       cp
#> 3 -3.696
#> # 2021-07-13 05:09:30 UTC: Iteration 2 of 2
#> # experimentsUsedSoFar: 15
#> # remainingBudget: 27
#> # currentBudget: 27
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          6|          4|          6|   0.2734375000|          4|00:00:00|   NA|  NA|    NA|
#> |x|          3|          4|          3|   0.2519531250|          7|00:00:00|-1.00|0.00|1.5000|
#> |x|          2|          4|          4|   0.2513020833|         10|00:00:00|-0.33|0.11|1.0000|
#> |x|          5|          4|          4|   0.2402343750|         13|00:00:00|-0.02|0.24|0.7893|
#> |=|          1|          4|          4|   0.2429687500|         16|00:00:00|+0.11|0.29|0.6560|
#> |=|          4|          4|          4|   0.2454427083|         19|00:00:00|-0.09|0.09|0.8005|
#> |=|          7|          4|          4|   0.2516741071|         23|00:00:00|-0.15|0.01|0.8571|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           4    mean value:    0.2516741071
#> Description of the best-so-far configuration:
#>   .ID.      cp .PARENT.
#> 4    4 -2.7744        3
#> 
#> # 2021-07-13 05:09:33 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        cp
#> 4 -2.7744
#> 3 -3.6960
#> # 2021-07-13 05:09:33 UTC: Stopped because there is not enough budget left to race more than the minimum (2)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 3
#> # nbIterations: 3
#> # experimentsUsedSoFar: 38
#> # timeUsed: 0
#> # remainingBudget: 4
#> # currentBudget: 4
#> # number of elites: 2
#> # nbConfigurations: 2

# best performing hyperparameter configuration
instance$result

#>         cp id_configuration learner_param_vals  x_domain classif.ce
#> 1: -2.7744                4          <list[2]> <list[1]>  0.2516741

# all evaluated hyperparameter configuration
as.data.table(instance$archive)

#>          cp classif.ce  x_domain_cp runtime_learners           timestamp
#>  1: -7.2126  0.2617188 0.0007372378            0.010 2021-07-13 05:09:28
#>  2: -5.1415  0.2617188 0.0058489098            0.010 2021-07-13 05:09:29
#>  3: -3.6960  0.2578125 0.0248226186            0.009 2021-07-13 05:09:29
#>  4: -7.2126  0.2851562 0.0007372378            0.011 2021-07-13 05:09:29
#>  5: -5.1415  0.2773438 0.0058489098            0.009 2021-07-13 05:09:29
#>  6: -3.6960  0.2734375 0.0248226186            0.012 2021-07-13 05:09:29
#>  7: -7.2126  0.2695312 0.0007372378            0.010 2021-07-13 05:09:29
#>  8: -5.1415  0.2695312 0.0058489098            0.011 2021-07-13 05:09:29
#>  9: -3.6960  0.2226562 0.0248226186            0.012 2021-07-13 05:09:29
#> 10: -7.2126  0.2851562 0.0007372378            0.011 2021-07-13 05:09:30
#> 11: -5.1415  0.2656250 0.0058489098            0.014 2021-07-13 05:09:30
#> 12: -3.6960  0.2578125 0.0248226186            0.017 2021-07-13 05:09:30
#> 13: -7.2126  0.2539062 0.0007372378            0.009 2021-07-13 05:09:30
#> 14: -5.1415  0.2343750 0.0058489098            0.010 2021-07-13 05:09:30
#> 15: -3.6960  0.2070312 0.0248226186            0.011 2021-07-13 05:09:30
#> 16: -3.6960  0.2812500 0.0248226186            0.010 2021-07-13 05:09:30
#> 17: -2.7744  0.2812500 0.0623868976            0.011 2021-07-13 05:09:30
#> 18: -4.0180  0.2773438 0.0179889068            0.010 2021-07-13 05:09:30
#> 19: -4.0939  0.2734375 0.0166740777            0.011 2021-07-13 05:09:31
#> 20: -2.7744  0.2226562 0.0623868976            0.010 2021-07-13 05:09:31
#> 21: -4.0180  0.2421875 0.0179889068            0.011 2021-07-13 05:09:31
#> 22: -4.0939  0.2617188 0.0166740777            0.011 2021-07-13 05:09:31
#> 23: -2.7744  0.2500000 0.0623868976            0.010 2021-07-13 05:09:31
#> 24: -4.0180  0.2773438 0.0179889068            0.009 2021-07-13 05:09:31
#> 25: -4.0939  0.2773438 0.0166740777            0.009 2021-07-13 05:09:31
#> 26: -2.7744  0.2070312 0.0623868976            0.011 2021-07-13 05:09:31
#> 27: -4.0180  0.2343750 0.0179889068            0.010 2021-07-13 05:09:31
#> 28: -4.0939  0.2343750 0.0166740777            0.011 2021-07-13 05:09:32
#> 29: -2.7744  0.2539062 0.0623868976            0.009 2021-07-13 05:09:32
#> 30: -4.0180  0.2578125 0.0179889068            0.011 2021-07-13 05:09:32
#> 31: -4.0939  0.2578125 0.0166740777            0.010 2021-07-13 05:09:32
#> 32: -2.7744  0.2578125 0.0623868976            0.010 2021-07-13 05:09:32
#> 33: -4.0180  0.2343750 0.0179889068            0.010 2021-07-13 05:09:32
#> 34: -4.0939  0.2343750 0.0166740777            0.011 2021-07-13 05:09:32
#> 35: -3.6960  0.2656250 0.0248226186            0.011 2021-07-13 05:09:32
#> 36: -2.7744  0.2890625 0.0623868976            0.010 2021-07-13 05:09:33
#> 37: -4.0180  0.2617188 0.0179889068            0.010 2021-07-13 05:09:33
#> 38: -4.0939  0.2617188 0.0166740777            0.012 2021-07-13 05:09:33
#>          cp classif.ce  x_domain_cp runtime_learners           timestamp
#>     batch_nr id_configuration id_instance      resample_result
#>  1:        1                1           8 <ResampleResult[20]>
#>  2:        2                2           8 <ResampleResult[20]>
#>  3:        3                3           8 <ResampleResult[20]>
#>  4:        4                1           2 <ResampleResult[20]>
#>  5:        5                2           2 <ResampleResult[20]>
#>  6:        6                3           2 <ResampleResult[20]>
#>  7:        7                1           3 <ResampleResult[20]>
#>  8:        8                2           3 <ResampleResult[20]>
#>  9:        9                3           3 <ResampleResult[20]>
#> 10:       10                1           5 <ResampleResult[20]>
#> 11:       11                2           5 <ResampleResult[20]>
#> 12:       12                3           5 <ResampleResult[20]>
#> 13:       13                1           6 <ResampleResult[20]>
#> 14:       14                2           6 <ResampleResult[20]>
#> 15:       15                3           6 <ResampleResult[20]>
#> 16:       16                3           1 <ResampleResult[20]>
#> 17:       17                4           1 <ResampleResult[20]>
#> 18:       18                5           1 <ResampleResult[20]>
#> 19:       19                6           1 <ResampleResult[20]>
#> 20:       20                4           3 <ResampleResult[20]>
#> 21:       21                5           3 <ResampleResult[20]>
#> 22:       22                6           3 <ResampleResult[20]>
#> 23:       23                4           2 <ResampleResult[20]>
#> 24:       24                5           2 <ResampleResult[20]>
#> 25:       25                6           2 <ResampleResult[20]>
#> 26:       26                4           6 <ResampleResult[20]>
#> 27:       27                5           6 <ResampleResult[20]>
#> 28:       28                6           6 <ResampleResult[20]>
#> 29:       29                4           8 <ResampleResult[20]>
#> 30:       30                5           8 <ResampleResult[20]>
#> 31:       31                6           8 <ResampleResult[20]>
#> 32:       32                4           5 <ResampleResult[20]>
#> 33:       33                5           5 <ResampleResult[20]>
#> 34:       34                6           5 <ResampleResult[20]>
#> 35:       35                3           7 <ResampleResult[20]>
#> 36:       36                4           7 <ResampleResult[20]>
#> 37:       37                5           7 <ResampleResult[20]>
#> 38:       38                6           7 <ResampleResult[20]>
#>     batch_nr id_configuration id_instance      resample_result

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(task)