Tuning via Iterated Racing.

TunerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi: 10.1016/j.orp.2016.09.002 .

Parameters

n_instances

integer(1)
Number of resampling instances.

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorRunTime or TerminatorEvals instead. Other terminators do not work with TunerIrace. We substract 5 seconds from the TerminatorRunTime budget for stability reasons.

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

TunerIrace$new()
mlr_tuners$get("irace")
tnr("irace")

See also

Super class

mlr3tuning::Tuner -> TunerIrace

Methods

Public methods

TunerIrace$new()

TunerIrace$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$optimize()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

instance = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE)),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = trm("evals", n_evals = 42)
)
tuner = tnr("irace")

# optimize hyperparameter
# modifies the instance by reference
tuner$optimize(instance)

#> # 2021-06-16 04:40:59 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 2
#> # minNbSurvival: 2
#> # nbParameters: 1
#> # seed: 1555455653
#> # confidence level: 0.95
#> # budget: 42
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2021-06-16 04:40:59 UTC: Iteration 1 of 2
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 42
#> # currentBudget: 21
#> # nbConfigurations: 3
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|          3|          1|  0.08000000000|          3|00:00:00|   NA|  NA|    NA|
#> |x|          2|          3|          1|  0.08000000000|          6|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|          3|          1|  0.07333333333|          9|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          3|          1|  0.07000000000|         12|00:00:00|+1.00|1.00|0.0000|
#> |=|          5|          3|          1|  0.06400000000|         15|00:00:00|+1.00|1.00|0.0000|
#> |=|          6|          3|          1|  0.05666666667|         18|00:00:00|+1.00|1.00|0.0000|
#> |=|          7|          3|          1|  0.05714285714|         21|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           1    mean value:   0.05714285714
#> Description of the best-so-far configuration:
#>   .ID.      cp .PARENT.
#> 1    1 -5.7461       NA
#> 
#> # 2021-06-16 04:41:01 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        cp
#> 1 -5.7461
#> 2 -9.0054
#> # 2021-06-16 04:41:01 UTC: Iteration 2 of 2
#> # experimentsUsedSoFar: 21
#> # remainingBudget: 21
#> # currentBudget: 21
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          8|          4|          1|   0.1000000000|          4|00:00:00|   NA|  NA|    NA|
#> |x|          7|          4|          1|  0.08000000000|          6|00:00:00|+1.00|1.00|0.0000|
#> |x|          6|          4|          1|  0.06000000000|          8|00:00:00|+1.00|1.00|0.0000|
#> |x|          5|          4|          1|  0.05500000000|         10|00:00:00|+1.00|1.00|0.0000|
#> |=|          2|          4|          1|  0.06000000000|         12|00:00:00|+1.00|1.00|0.0000|
#> |=|          1|          4|          1|  0.06333333333|         14|00:00:00|+1.00|1.00|0.0000|
#> |=|          4|          4|          1|  0.06285714286|         16|00:00:00|+1.00|1.00|0.0000|
#> |=|          3|          4|          1|  0.06250000000|         18|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           1    mean value:   0.06250000000
#> Description of the best-so-far configuration:
#>   .ID.      cp .PARENT.
#> 1    1 -5.7461       NA
#> 
#> # 2021-06-16 04:41:03 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        cp
#> 1 -5.7461
#> 2 -9.0054
#> # 2021-06-16 04:41:03 UTC: Stopped because there is not enough budget left to race more than the minimum (2)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 3
#> # nbIterations: 3
#> # experimentsUsedSoFar: 39
#> # timeUsed: 0
#> # remainingBudget: 3
#> # currentBudget: 3
#> # number of elites: 2
#> # nbConfigurations: 2

#>         cp id_configuration learner_param_vals  x_domain classif.ce
#> 1: -5.7461                1          <list[2]> <list[1]>     0.0625

# returns best configuration and best performance
instance$result

#>         cp id_configuration learner_param_vals  x_domain classif.ce
#> 1: -5.7461                1          <list[2]> <list[1]>     0.0625

# allows access of data.table of full path of all evaluations
instance$archive

#> <ArchiveTuning>
#>       cp classif.ce runtime_learners           timestamp batch_nr
#>  1: -9.1       0.08            0.008 2021-06-16 04:41:00        6
#>  2: -9.1       0.06            0.007 2021-06-16 04:41:00       12
#>  3: -9.0       0.06            0.007 2021-06-16 04:41:01       20
#>  4: -5.7       0.04            0.007 2021-06-16 04:41:00       13
#>  5: -5.7       0.06            0.007 2021-06-16 04:41:00        7
#>  6: -5.4       0.06            0.006 2021-06-16 04:41:03       37
#>  7: -5.4       0.10            0.006 2021-06-16 04:41:02       25
#>  8: -7.1       0.04            0.007 2021-06-16 04:41:02       30
#>  9: -7.1       0.08            0.007 2021-06-16 04:41:02       32
#> 10: -5.7       0.02            0.007 2021-06-16 04:41:01       16
#> 11: -5.4       0.06            0.008 2021-06-16 04:41:03       39
#> 12: -5.4       0.04            0.007 2021-06-16 04:41:02       31
#> 13: -5.4       0.08            0.008 2021-06-16 04:41:02       33
#> 14: -9.0       0.02            0.007 2021-06-16 04:41:01       17
#> 15: -5.4       0.02            0.007 2021-06-16 04:41:02       29
#> 16: -7.1       0.08            0.007 2021-06-16 04:41:02       34
#> 17: -9.1       0.02            0.008 2021-06-16 04:41:01       18
#> 18: -5.7       0.10            0.007 2021-06-16 04:41:01       22
#> 19: -9.1       0.08            0.007 2021-06-16 04:40:59        3
#> 20: -5.7       0.08            0.009 2021-06-16 04:40:59        4
#> 21: -9.1       0.04            0.006 2021-06-16 04:41:01       15
#> 22: -9.0       0.10            0.007 2021-06-16 04:41:01       23
#> 23: -7.1       0.10            0.007 2021-06-16 04:41:01       24
#> 24: -9.0       0.08            0.007 2021-06-16 04:41:00        5
#> 25: -9.0       0.08            0.007 2021-06-16 04:40:59        2
#> 26: -7.1       0.02            0.007 2021-06-16 04:41:02       28
#> 27: -9.0       0.04            0.007 2021-06-16 04:41:00       14
#> 28: -7.1       0.06            0.007 2021-06-16 04:41:03       36
#> 29: -9.0       0.06            0.007 2021-06-16 04:41:00        8
#> 30: -5.7       0.06            0.007 2021-06-16 04:41:00       10
#> 31: -9.1       0.06            0.007 2021-06-16 04:41:01       21
#> 32: -7.1       0.06            0.008 2021-06-16 04:41:03       38
#> 33: -9.1       0.06            0.007 2021-06-16 04:41:00        9
#> 34: -7.1       0.06            0.007 2021-06-16 04:41:02       26
#> 35: -5.7       0.06            0.007 2021-06-16 04:41:01       19
#> 36: -5.7       0.08            0.008 2021-06-16 04:40:59        1
#> 37: -9.0       0.06            0.007 2021-06-16 04:41:00       11
#> 38: -5.4       0.06            0.007 2021-06-16 04:41:02       27
#> 39: -5.4       0.08            0.011 2021-06-16 04:41:02       35
#>       cp classif.ce runtime_learners           timestamp batch_nr
#>     id_configuration id_instance      resample_result
#>  1:                3          10 <ResampleResult[20]>
#>  2:                3           2 <ResampleResult[20]>
#>  3:                2           4 <ResampleResult[20]>
#>  4:                1           8 <ResampleResult[20]>
#>  5:                1           1 <ResampleResult[20]>
#>  6:                5           2 <ResampleResult[20]>
#>  7:                5           5 <ResampleResult[20]>
#>  8:                4           8 <ResampleResult[20]>
#>  9:                4          10 <ResampleResult[20]>
#> 10:                1           7 <ResampleResult[20]>
#> 11:                5           1 <ResampleResult[20]>
#> 12:                5           8 <ResampleResult[20]>
#> 13:                5          10 <ResampleResult[20]>
#> 14:                2           7 <ResampleResult[20]>
#> 15:                5           7 <ResampleResult[20]>
#> 16:                4           3 <ResampleResult[20]>
#> 17:                3           7 <ResampleResult[20]>
#> 18:                1           5 <ResampleResult[20]>
#> 19:                3           3 <ResampleResult[20]>
#> 20:                1          10 <ResampleResult[20]>
#> 21:                3           8 <ResampleResult[20]>
#> 22:                2           5 <ResampleResult[20]>
#> 23:                4           5 <ResampleResult[20]>
#> 24:                2          10 <ResampleResult[20]>
#> 25:                2           3 <ResampleResult[20]>
#> 26:                4           7 <ResampleResult[20]>
#> 27:                2           8 <ResampleResult[20]>
#> 28:                4           2 <ResampleResult[20]>
#> 29:                2           1 <ResampleResult[20]>
#> 30:                1           2 <ResampleResult[20]>
#> 31:                3           4 <ResampleResult[20]>
#> 32:                4           1 <ResampleResult[20]>
#> 33:                3           1 <ResampleResult[20]>
#> 34:                4           4 <ResampleResult[20]>
#> 35:                1           4 <ResampleResult[20]>
#> 36:                1           3 <ResampleResult[20]>
#> 37:                2           2 <ResampleResult[20]>
#> 38:                5           4 <ResampleResult[20]>
#> 39:                5           3 <ResampleResult[20]>
#>     id_configuration id_instance      resample_result