Tuning via Iterated Racing.

TunerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi: 10.1016/j.orp.2016.09.002 .

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

TunerIrace$new()
mlr_tuners$get("irace")
tnr("irace")

Parameters

n_instances

integer(1)
Number of resampling instances.

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with TunerIrace.

Archive

The ArchiveTuning holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies resampling instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The tuning result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerIrace

Methods

Public methods

TunerIrace$new()

TunerIrace$optimize()

TunerIrace$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerIrace$new()

Method optimize()

Performs the tuning on a TuningInstanceSingleCrit until termination. The single evaluations and the final results will be written into the ArchiveTuning that resides in the TuningInstanceSingleCrit. The final result is returned.

Usage

TunerIrace$optimize(inst)

Arguments

inst

(TuningInstanceSingleCrit).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner and set search space
learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE))

# hyperparameter tuning on the pima indians diabetes data set
instance = tune(
  method = "irace",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 42
)
#> # 2021-11-05 04:25:58 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 2
#> # minNbSurvival: 2
#> # nbParameters: 1
#> # seed: 983633176
#> # confidence level: 0.95
#> # budget: 42
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2021-11-05 04:25:58 UTC: Iteration 1 of 2
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 42
#> # currentBudget: 21
#> # nbConfigurations: 3
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|          3|          3|   0.2578125000|          3|00:00:00|   NA|  NA|    NA|
#> |x|          2|          3|          3|   0.2656250000|          6|00:00:00|+0.86|0.93|0.1627|
#> |x|          3|          3|          3|   0.2513020833|          9|00:00:00|+0.90|0.93|0.1085|
#> |x|          4|          3|          3|   0.2529296875|         12|00:00:00|+0.90|0.93|0.1165|
#> |-|          5|          1|          3|   0.2437500000|         15|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           3    mean value:    0.2437500000
#> Description of the best-so-far configuration:
#>   .ID.     cp .PARENT.
#> 3    3 -3.696       NA
#> 
#> # 2021-11-05 04:25:59 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>       cp
#> 3 -3.696
#> # 2021-11-05 04:25:59 UTC: Iteration 2 of 2
#> # experimentsUsedSoFar: 15
#> # remainingBudget: 27
#> # currentBudget: 27
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          6|          4|          3|   0.2812500000|          4|00:00:00|   NA|  NA|    NA|
#> |x|          1|          4|          3|   0.2695312500|          7|00:00:00|+0.00|0.50|0.3750|
#> |x|          5|          4|          3|   0.2486979167|         10|00:00:00|+0.50|0.67|0.2500|
#> |x|          4|          4|          3|   0.2509765625|         13|00:00:00|+0.33|0.50|0.2500|
#> |=|          2|          4|          4|   0.2507812500|         16|00:00:00|+0.06|0.24|0.4250|
#> |=|          3|          4|          4|   0.2460937500|         19|00:00:00|+0.04|0.20|0.3583|
#> |=|          7|          4|          3|   0.2522321429|         23|00:00:00|+0.00|0.14|0.4286|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           3    mean value:    0.2522321429
#> Description of the best-so-far configuration:
#>   .ID.     cp .PARENT.
#> 3    3 -3.696       NA
#> 
#> # 2021-11-05 04:26:01 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>        cp
#> 3 -3.6960
#> 4 -3.2257
#> # 2021-11-05 04:26:01 UTC: Stopped because there is not enough budget left to race more than the minimum (2)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 3
#> # nbIterations: 3
#> # experimentsUsedSoFar: 38
#> # timeUsed: 0
#> # remainingBudget: 4
#> # currentBudget: 4
#> # number of elites: 2
#> # nbConfigurations: 2

# best performing hyperparameter configuration
instance$result
#>        cp configuration learner_param_vals  x_domain classif.ce
#> 1: -3.696             3          <list[2]> <list[1]>  0.2522321

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>          cp classif.ce  x_domain_cp runtime_learners           timestamp
#>  1: -7.2126  0.2617188 0.0007372378            0.016 2021-11-05 04:25:58
#>  2: -5.1415  0.2617188 0.0058489098            0.016 2021-11-05 04:25:58
#>  3: -3.6960  0.2578125 0.0248226186            0.018 2021-11-05 04:25:58
#>  4: -7.2126  0.2851562 0.0007372378            0.016 2021-11-05 04:25:58
#>  5: -5.1415  0.2773438 0.0058489098            0.016 2021-11-05 04:25:58
#>  6: -3.6960  0.2734375 0.0248226186            0.018 2021-11-05 04:25:58
#>  7: -7.2126  0.2695312 0.0007372378            0.016 2021-11-05 04:25:58
#>  8: -5.1415  0.2695312 0.0058489098            0.018 2021-11-05 04:25:58
#>  9: -3.6960  0.2226562 0.0248226186            0.015 2021-11-05 04:25:58
#> 10: -7.2126  0.2851562 0.0007372378            0.016 2021-11-05 04:25:59
#> 11: -5.1415  0.2656250 0.0058489098            0.016 2021-11-05 04:25:59
#> 12: -3.6960  0.2578125 0.0248226186            0.020 2021-11-05 04:25:59
#> 13: -7.2126  0.2539062 0.0007372378            0.015 2021-11-05 04:25:59
#> 14: -5.1415  0.2343750 0.0058489098            0.016 2021-11-05 04:25:59
#> 15: -3.6960  0.2070312 0.0248226186            0.015 2021-11-05 04:25:59
#> 16: -3.6960  0.2812500 0.0248226186            0.016 2021-11-05 04:25:59
#> 17: -3.2257  0.2812500 0.0397279623            0.014 2021-11-05 04:25:59
#> 18: -2.5684  0.2812500 0.0766581003            0.015 2021-11-05 04:25:59
#> 19: -3.3355  0.2812500 0.0355967834            0.013 2021-11-05 04:25:59
#> 20: -3.2257  0.2578125 0.0397279623            0.016 2021-11-05 04:26:00
#> 21: -2.5684  0.3085938 0.0766581003            0.019 2021-11-05 04:26:00
#> 22: -3.3355  0.2578125 0.0355967834            0.015 2021-11-05 04:26:00
#> 23: -3.2257  0.2070312 0.0397279623            0.017 2021-11-05 04:26:00
#> 24: -2.5684  0.2656250 0.0766581003            0.016 2021-11-05 04:26:00
#> 25: -3.3355  0.2070312 0.0355967834            0.037 2021-11-05 04:26:00
#> 26: -3.2257  0.2578125 0.0397279623            0.016 2021-11-05 04:26:00
#> 27: -2.5684  0.2578125 0.0766581003            0.016 2021-11-05 04:26:00
#> 28: -3.3355  0.2578125 0.0355967834            0.016 2021-11-05 04:26:00
#> 29: -3.2257  0.2500000 0.0397279623            0.016 2021-11-05 04:26:01
#> 30: -2.5684  0.2500000 0.0766581003            0.020 2021-11-05 04:26:01
#> 31: -3.3355  0.2500000 0.0355967834            0.062 2021-11-05 04:26:01
#> 32: -3.2257  0.2226562 0.0397279623            0.017 2021-11-05 04:26:01
#> 33: -2.5684  0.2226562 0.0766581003            0.014 2021-11-05 04:26:01
#> 34: -3.3355  0.2226562 0.0355967834            0.015 2021-11-05 04:26:01
#> 35: -3.6960  0.2656250 0.0248226186            0.020 2021-11-05 04:26:01
#> 36: -3.2257  0.2890625 0.0397279623            0.016 2021-11-05 04:26:01
#> 37: -2.5684  0.2890625 0.0766581003            0.018 2021-11-05 04:26:01
#> 38: -3.3355  0.2890625 0.0355967834            0.019 2021-11-05 04:26:01
#>          cp classif.ce  x_domain_cp runtime_learners           timestamp
#>     batch_nr race step instance configuration      resample_result
#>  1:        1    1    1        8             1 <ResampleResult[22]>
#>  2:        1    1    1        8             2 <ResampleResult[22]>
#>  3:        1    1    1        8             3 <ResampleResult[22]>
#>  4:        2    1    1        2             1 <ResampleResult[22]>
#>  5:        2    1    1        2             2 <ResampleResult[22]>
#>  6:        2    1    1        2             3 <ResampleResult[22]>
#>  7:        3    1    1        3             1 <ResampleResult[22]>
#>  8:        3    1    1        3             2 <ResampleResult[22]>
#>  9:        3    1    1        3             3 <ResampleResult[22]>
#> 10:        4    1    1        5             1 <ResampleResult[22]>
#> 11:        4    1    1        5             2 <ResampleResult[22]>
#> 12:        4    1    1        5             3 <ResampleResult[22]>
#> 13:        5    1    1        6             1 <ResampleResult[22]>
#> 14:        5    1    1        6             2 <ResampleResult[22]>
#> 15:        5    1    1        6             3 <ResampleResult[22]>
#> 16:        6    2    1        1             3 <ResampleResult[22]>
#> 17:        6    2    1        1             4 <ResampleResult[22]>
#> 18:        6    2    1        1             5 <ResampleResult[22]>
#> 19:        6    2    1        1             6 <ResampleResult[22]>
#> 20:        7    2    1        8             4 <ResampleResult[22]>
#> 21:        7    2    1        8             5 <ResampleResult[22]>
#> 22:        7    2    1        8             6 <ResampleResult[22]>
#> 23:        8    2    1        6             4 <ResampleResult[22]>
#> 24:        8    2    1        6             5 <ResampleResult[22]>
#> 25:        8    2    1        6             6 <ResampleResult[22]>
#> 26:        9    2    1        5             4 <ResampleResult[22]>
#> 27:        9    2    1        5             5 <ResampleResult[22]>
#> 28:        9    2    1        5             6 <ResampleResult[22]>
#> 29:       10    2    1        2             4 <ResampleResult[22]>
#> 30:       10    2    1        2             5 <ResampleResult[22]>
#> 31:       10    2    1        2             6 <ResampleResult[22]>
#> 32:       11    2    1        3             4 <ResampleResult[22]>
#> 33:       11    2    1        3             5 <ResampleResult[22]>
#> 34:       11    2    1        3             6 <ResampleResult[22]>
#> 35:       12    2    1        7             3 <ResampleResult[22]>
#> 36:       12    2    1        7             4 <ResampleResult[22]>
#> 37:       12    2    1        7             5 <ResampleResult[22]>
#> 38:       12    2    1        7             6 <ResampleResult[22]>
#>     batch_nr race step instance configuration      resample_result

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(task)