Tuning via Iterated Racing.

TunerIrace class that implements iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi: 10.1016/j.orp.2016.09.002 .

Parameters

n_instances

integer(1)
Number of resampling instances.

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorRunTime or TerminatorEvals instead. Other terminators do not work with TunerIrace. We substract 5 seconds from the TerminatorRunTime budget for stability reasons.

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

TunerIrace$new()
mlr_tuners$get("irace")
tnr("irace")

See also

Super class

mlr3tuning::Tuner -> TunerIrace

Methods

Public methods

TunerIrace$new()

TunerIrace$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$optimize()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerIrace$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(paradox)
search_space = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1)
))
terminator = trm("evals", n_evals = 42)
instance = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  search_space = search_space,
  terminator = terminator
)
tt = tnr("irace")

# modifies the instance by reference
tt$optimize(instance)

#> # 2021-03-30 04:23:51 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 2
#> # minNbSurvival: 2
#> # nbParameters: 1
#> # seed: 962396134
#> # confidence level: 0.95
#> # budget: 42
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2021-03-30 04:23:51 UTC: Iteration 1 of 2
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 42
#> # currentBudget: 21
#> # nbConfigurations: 3
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          1|          3|          1|  0.02000000000|          3|00:00:00|   NA|  NA|    NA|
#> |x|          2|          3|          1|  0.04000000000|          6|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|          3|          1|  0.03333333333|          9|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          3|          1|  0.03500000000|         12|00:00:00|+1.00|1.00|0.0000|
#> |=|          5|          3|          1|  0.03600000000|         15|00:00:00|+1.00|1.00|0.0000|
#> |=|          6|          3|          1|  0.04666666667|         18|00:00:00|+1.00|1.00|0.0000|
#> |=|          7|          3|          1|  0.04857142857|         21|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           1    mean value:   0.04857142857
#> Description of the best-so-far configuration:
#>   .ID.     cp .PARENT.
#> 1    1 0.0519       NA
#> 
#> # 2021-03-30 04:23:53 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>       cp
#> 1 0.0519
#> 2 0.0368
#> # 2021-03-30 04:23:53 UTC: Iteration 2 of 2
#> # experimentsUsedSoFar: 21
#> # remainingBudget: 21
#> # currentBudget: 21
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|      Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> |x|          8|          4|          1|  0.06000000000|          4|00:00:00|   NA|  NA|    NA|
#> |x|          3|          4|          1|  0.04000000000|          6|00:00:00|+1.00|1.00|0.0000|
#> |x|          7|          4|          1|  0.04666666667|          8|00:00:00|+1.00|1.00|0.0000|
#> |x|          1|          4|          1|  0.04000000000|         10|00:00:00|+1.00|1.00|0.0000|
#> |=|          4|          4|          1|  0.04000000000|         12|00:00:00|+1.00|1.00|0.0000|
#> |=|          5|          4|          1|  0.04000000000|         14|00:00:00|+1.00|1.00|0.0000|
#> |=|          6|          4|          1|  0.04857142857|         16|00:00:00|+1.00|1.00|0.0000|
#> |=|          2|          4|          1|  0.05000000000|         18|00:00:00|+1.00|1.00|0.0000|
#> +-+-----------+-----------+-----------+---------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           1    mean value:   0.05000000000
#> Description of the best-so-far configuration:
#>   .ID.     cp .PARENT.
#> 1    1 0.0519       NA
#> 
#> # 2021-03-30 04:23:55 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>       cp
#> 1 0.0519
#> 2 0.0368
#> # 2021-03-30 04:23:55 UTC: Stopped because there is not enough budget left to race more than the minimum (2)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 3
#> # nbIterations: 3
#> # experimentsUsedSoFar: 39
#> # timeUsed: 0
#> # remainingBudget: 3
#> # currentBudget: 3
#> # number of elites: 2
#> # nbConfigurations: 2

#>        cp id_configuration learner_param_vals  x_domain classif.ce
#> 1: 0.0519                1          <list[2]> <list[1]>       0.05

# returns best configuration and best performance
instance$result

#>        cp id_configuration learner_param_vals  x_domain classif.ce
#> 1: 0.0519                1          <list[2]> <list[1]>       0.05

# allows access of data.table of full path of all evaluations
instance$archive

#> <ArchiveTuning>
#>        cp id_configuration id_instance classif.ce           timestamp batch_nr
#>  1: 0.052                1           7       0.02 2021-03-30 04:23:51        1
#>  2: 0.037                2           7       0.02 2021-03-30 04:23:51        2
#>  3: 0.040                3           7       0.02 2021-03-30 04:23:51        3
#>  4: 0.052                1           2       0.06 2021-03-30 04:23:51        4
#>  5: 0.037                2           2       0.06 2021-03-30 04:23:51        5
#>  6: 0.040                3           2       0.06 2021-03-30 04:23:52        6
#>  7: 0.052                1           4       0.02 2021-03-30 04:23:52        7
#>  8: 0.037                2           4       0.02 2021-03-30 04:23:52        8
#>  9: 0.040                3           4       0.02 2021-03-30 04:23:52        9
#> 10: 0.052                1           9       0.04 2021-03-30 04:23:52       10
#> 11: 0.037                2           9       0.04 2021-03-30 04:23:52       11
#> 12: 0.040                3           9       0.04 2021-03-30 04:23:52       12
#> 13: 0.052                1           8       0.04 2021-03-30 04:23:52       13
#> 14: 0.037                2           8       0.04 2021-03-30 04:23:52       14
#> 15: 0.040                3           8       0.04 2021-03-30 04:23:53       15
#> 16: 0.052                1           6       0.10 2021-03-30 04:23:53       16
#> 17: 0.037                2           6       0.10 2021-03-30 04:23:53       17
#> 18: 0.040                3           6       0.10 2021-03-30 04:23:53       18
#> 19: 0.052                1           5       0.06 2021-03-30 04:23:53       19
#> 20: 0.037                2           5       0.06 2021-03-30 04:23:53       20
#> 21: 0.040                3           5       0.06 2021-03-30 04:23:53       21
#> 22: 0.052                1           1       0.06 2021-03-30 04:23:53       22
#> 23: 0.037                2           1       0.06 2021-03-30 04:23:53       23
#> 24: 0.038                4           1       0.06 2021-03-30 04:23:54       24
#> 25: 0.051                5           1       0.06 2021-03-30 04:23:54       25
#> 26: 0.038                4           4       0.02 2021-03-30 04:23:54       26
#> 27: 0.051                5           4       0.02 2021-03-30 04:23:54       27
#> 28: 0.038                4           5       0.06 2021-03-30 04:23:54       28
#> 29: 0.051                5           5       0.06 2021-03-30 04:23:54       29
#> 30: 0.038                4           7       0.02 2021-03-30 04:23:54       30
#> 31: 0.051                5           7       0.02 2021-03-30 04:23:54       31
#> 32: 0.038                4           9       0.04 2021-03-30 04:23:54       32
#> 33: 0.051                5           9       0.04 2021-03-30 04:23:55       33
#> 34: 0.038                4           8       0.04 2021-03-30 04:23:55       34
#> 35: 0.051                5           8       0.04 2021-03-30 04:23:55       35
#> 36: 0.038                4           6       0.10 2021-03-30 04:23:55       36
#> 37: 0.051                5           6       0.10 2021-03-30 04:23:55       37
#> 38: 0.038                4           2       0.06 2021-03-30 04:23:55       38
#> 39: 0.051                5           2       0.06 2021-03-30 04:23:55       39
#>        cp id_configuration id_instance classif.ce           timestamp batch_nr