Hyperparameter Tuning with Iterated Racing.

Subclass for iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi:10.1016/j.orp.2016.09.002 .

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("irace")

Control Parameters

n_instances

integer(1)
Number of resampling instances.

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with TunerIrace.

Archive

The ArchiveBatchTuning holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies resampling instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The tuning result (instance$result) is the best-performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerBatchIrace which can be applied on any black box optimization problem. See also the documentation of bbotk.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

An overview of all tuners can be found on our website.

Learn more about tuners.

The gallery features a collection of case studies and demos about optimization.

Use the Hyperband optimizer with different budget parameters.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerBatch -> mlr3tuning::TunerBatchFromOptimizerBatch -> TunerBatchIrace

Methods

Public methods

TunerBatchIrace$new()

TunerBatchIrace$optimize()

TunerBatchIrace$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerBatchIrace$new()

Method optimize()

Performs the tuning on a TuningInstanceBatchSingleCrit until termination. The single evaluations and the final results will be written into the ArchiveBatchTuning that resides in the TuningInstanceBatchSingleCrit. The final result is returned.

Usage

TunerBatchIrace$optimize(inst)

Arguments

inst

(TuningInstanceBatchSingleCrit).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerBatchIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner and set search space
learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE))
# \donttest{
# hyperparameter tuning on the pima indians diabetes data set
instance = tune(
  tuner = tnr("irace"),
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 42
)
#> # 2024-06-20 10:09:00 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 2
#> # minNbSurvival: 2
#> # nbParameters: 1
#> # seed: 1855097766
#> # confidence level: 0.95
#> # budget: 42
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2024-06-20 10:09:00 UTC: Iteration 1 of 2
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 42
#> # currentBudget: 21
#> # nbConfigurations: 3
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          1|          3|          3|    0.2812500000|          3|00:00:00|   NA|  NA|    NA|
#> |x|          2|          3|          3|    0.2636718750|          6|00:00:00|+1.00|1.00|0.0000|
#> |x|          3|          3|          3|    0.2460937500|          9|00:00:00|+1.00|1.00|0.0000|
#> |x|          4|          3|          3|    0.2382812500|         12|00:00:00|+1.00|1.00|0.0000|
#> |-|          5|          1|          3|    0.2367187500|         15|00:00:00|   NA|  NA|    NA|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           3    mean value:     0.2367187500
#> Description of the best-so-far configuration:
#>   .ID.               cp .PARENT.
#> 3    3 -2.7229877489945       NA
#> 
#> # 2024-06-20 10:09:01 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                 cp
#> 3 -2.7229877489945
#> # 2024-06-20 10:09:01 UTC: Iteration 2 of 2
#> # experimentsUsedSoFar: 15
#> # remainingBudget: 27
#> # currentBudget: 27
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          6|          4|          4|    0.2734375000|          4|00:00:00|   NA|  NA|    NA|
#> |x|          1|          4|          4|    0.2597656250|          7|00:00:00|+0.80|0.90|0.0217|
#> |x|          4|          4|          4|    0.2447916667|         10|00:00:00|+0.40|0.60|0.2572|
#> |x|          3|          4|          4|    0.2363281250|         13|00:00:00|+0.27|0.45|0.2536|
#> |=|          5|          4|          4|    0.2351562500|         16|00:00:00|+0.20|0.36|0.2272|
#> |=|          2|          4|          4|    0.2363281250|         19|00:00:00|+0.12|0.26|0.3681|
#> |=|          7|          4|          4|    0.2399553571|         23|00:00:00|-0.08|0.08|0.4776|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           4    mean value:     0.2399553571
#> Description of the best-so-far configuration:
#>   .ID.                cp .PARENT.
#> 4    4 -3.68087230803958        3
#> 
#> # 2024-06-20 10:09:02 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  cp
#> 4 -3.68087230803958
#> 5 -3.17982221206359
#> # 2024-06-20 10:09:02 UTC: Stopped because there is not enough budget left to race more than the minimum (2)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 3
#> # nbIterations: 3
#> # experimentsUsedSoFar: 38
#> # timeUsed: 0
#> # remainingBudget: 4
#> # currentBudget: 4
#> # number of elites: 2
#> # nbConfigurations: 2
#> # Total CPU user time: 1.784, CPU sys time: 0.009, Wall-clock time: 1.793

# best performing hyperparameter configuration
instance$result
#>           cp configuration learner_param_vals  x_domain classif.ce
#>        <num>         <num>             <list>    <list>      <num>
#> 1: -3.680872             4          <list[2]> <list[1]>  0.2399554

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp classif.ce  x_domain_cp runtime_learners           timestamp
#>         <num>      <num>        <num>            <num>              <POSc>
#>  1: -8.192526  0.3085938 0.0002767141            0.013 2024-06-20 10:09:01
#>  2: -8.612223  0.3085938 0.0001818692            0.013 2024-06-20 10:09:01
#>  3: -2.722988  0.2812500 0.0656782309            0.011 2024-06-20 10:09:01
#>  4: -8.192526  0.3593750 0.0002767141            0.032 2024-06-20 10:09:01
#>  5: -8.612223  0.3593750 0.0001818692            0.016 2024-06-20 10:09:01
#>  6: -2.722988  0.2460938 0.0656782309            0.013 2024-06-20 10:09:01
#>  7: -8.192526  0.2578125 0.0002767141            0.015 2024-06-20 10:09:01
#>  8: -8.612223  0.2578125 0.0001818692            0.013 2024-06-20 10:09:01
#>  9: -2.722988  0.2109375 0.0656782309            0.014 2024-06-20 10:09:01
#> 10: -8.192526  0.2617188 0.0002767141            0.013 2024-06-20 10:09:01
#> 11: -8.612223  0.2617188 0.0001818692            0.011 2024-06-20 10:09:01
#> 12: -2.722988  0.2148438 0.0656782309            0.013 2024-06-20 10:09:01
#> 13: -8.192526  0.2382812 0.0002767141            0.019 2024-06-20 10:09:01
#> 14: -8.612223  0.2382812 0.0001818692            0.013 2024-06-20 10:09:01
#> 15: -2.722988  0.2304688 0.0656782309            0.013 2024-06-20 10:09:01
#> 16: -2.722988  0.2890625 0.0656782309            0.014 2024-06-20 10:09:01
#> 17: -3.680872  0.2734375 0.0252009822            0.013 2024-06-20 10:09:01
#> 18: -3.179822  0.2734375 0.0415930492            0.059 2024-06-20 10:09:01
#> 19: -3.293965  0.2734375 0.0371064177            0.014 2024-06-20 10:09:01
#> 20: -3.680872  0.2460938 0.0252009822            0.011 2024-06-20 10:09:01
#> 21: -3.179822  0.2539062 0.0415930492            0.013 2024-06-20 10:09:01
#> 22: -3.293965  0.2539062 0.0371064177            0.011 2024-06-20 10:09:01
#> 23: -3.680872  0.2148438 0.0252009822            0.014 2024-06-20 10:09:02
#> 24: -3.179822  0.2148438 0.0415930492            0.012 2024-06-20 10:09:02
#> 25: -3.293965  0.2148438 0.0371064177            0.013 2024-06-20 10:09:02
#> 26: -3.680872  0.2109375 0.0252009822            0.013 2024-06-20 10:09:02
#> 27: -3.179822  0.2109375 0.0415930492            0.030 2024-06-20 10:09:02
#> 28: -3.293965  0.2109375 0.0371064177            0.017 2024-06-20 10:09:02
#> 29: -3.680872  0.2304688 0.0252009822            0.013 2024-06-20 10:09:02
#> 30: -3.179822  0.2304688 0.0415930492            0.013 2024-06-20 10:09:02
#> 31: -3.293965  0.2304688 0.0371064177            0.012 2024-06-20 10:09:02
#> 32: -3.680872  0.2421875 0.0252009822            0.014 2024-06-20 10:09:02
#> 33: -3.179822  0.2890625 0.0415930492            0.013 2024-06-20 10:09:02
#> 34: -3.293965  0.2890625 0.0371064177            0.013 2024-06-20 10:09:02
#> 35: -2.722988  0.2500000 0.0656782309            0.013 2024-06-20 10:09:02
#> 36: -3.680872  0.2617188 0.0252009822            0.028 2024-06-20 10:09:02
#> 37: -3.179822  0.2500000 0.0415930492            0.017 2024-06-20 10:09:02
#> 38: -3.293965  0.2500000 0.0371064177            0.013 2024-06-20 10:09:02
#>            cp classif.ce  x_domain_cp runtime_learners           timestamp
#>     batch_nr  race  step instance configuration warnings errors
#>        <int> <num> <int>    <int>         <num>    <int>  <int>
#>  1:        1     1     1       10             1        0      0
#>  2:        1     1     1       10             2        0      0
#>  3:        1     1     1       10             3        0      0
#>  4:        2     1     1        4             1        0      0
#>  5:        2     1     1        4             2        0      0
#>  6:        2     1     1        4             3        0      0
#>  7:        3     1     1        1             1        0      0
#>  8:        3     1     1        1             2        0      0
#>  9:        3     1     1        1             3        0      0
#> 10:        4     1     1        8             1        0      0
#> 11:        4     1     1        8             2        0      0
#> 12:        4     1     1        8             3        0      0
#> 13:        5     1     1        5             1        0      0
#> 14:        5     1     1        5             2        0      0
#> 15:        5     1     1        5             3        0      0
#> 16:        6     2     1        7             3        0      0
#> 17:        6     2     1        7             4        0      0
#> 18:        6     2     1        7             5        0      0
#> 19:        6     2     1        7             6        0      0
#> 20:        7     2     1       10             4        0      0
#> 21:        7     2     1       10             5        0      0
#> 22:        7     2     1       10             6        0      0
#> 23:        8     2     1        8             4        0      0
#> 24:        8     2     1        8             5        0      0
#> 25:        8     2     1        8             6        0      0
#> 26:        9     2     1        1             4        0      0
#> 27:        9     2     1        1             5        0      0
#> 28:        9     2     1        1             6        0      0
#> 29:       10     2     1        5             4        0      0
#> 30:       10     2     1        5             5        0      0
#> 31:       10     2     1        5             6        0      0
#> 32:       11     2     1        4             4        0      0
#> 33:       11     2     1        4             5        0      0
#> 34:       11     2     1        4             6        0      0
#> 35:       12     2     1        9             3        0      0
#> 36:       12     2     1        9             4        0      0
#> 37:       12     2     1        9             5        0      0
#> 38:       12     2     1        9             6        0      0
#>     batch_nr  race  step instance configuration warnings errors
#>      resample_result
#>               <list>
#>  1: <ResampleResult>
#>  2: <ResampleResult>
#>  3: <ResampleResult>
#>  4: <ResampleResult>
#>  5: <ResampleResult>
#>  6: <ResampleResult>
#>  7: <ResampleResult>
#>  8: <ResampleResult>
#>  9: <ResampleResult>
#> 10: <ResampleResult>
#> 11: <ResampleResult>
#> 12: <ResampleResult>
#> 13: <ResampleResult>
#> 14: <ResampleResult>
#> 15: <ResampleResult>
#> 16: <ResampleResult>
#> 17: <ResampleResult>
#> 18: <ResampleResult>
#> 19: <ResampleResult>
#> 20: <ResampleResult>
#> 21: <ResampleResult>
#> 22: <ResampleResult>
#> 23: <ResampleResult>
#> 24: <ResampleResult>
#> 25: <ResampleResult>
#> 26: <ResampleResult>
#> 27: <ResampleResult>
#> 28: <ResampleResult>
#> 29: <ResampleResult>
#> 30: <ResampleResult>
#> 31: <ResampleResult>
#> 32: <ResampleResult>
#> 33: <ResampleResult>
#> 34: <ResampleResult>
#> 35: <ResampleResult>
#> 36: <ResampleResult>
#> 37: <ResampleResult>
#> 38: <ResampleResult>
#>      resample_result

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(task)
# }