Hyperparameter Tuning with Iterated Racing.

Subclass for iterated racing. Calls irace::irace() from package irace.

Source

Lopez-Ibanez M, Dubois-Lacoste J, Caceres LP, Birattari M, Stuetzle T (2016). “The irace package: Iterated racing for automatic algorithm configuration.” Operations Research Perspectives, 3, 43--58. doi:10.1016/j.orp.2016.09.002 .

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("irace")

Control Parameters

n_instances

integer(1)
Number of resampling instances.

For the meaning of all other parameters, see irace::defaultScenario(). Note that we have removed all control parameters which refer to the termination of the algorithm. Use TerminatorEvals instead. Other terminators do not work with TunerIrace.

Archive

The ArchiveTuning holds the following additional columns:

"race" (integer(1))
Race iteration.

"step" (integer(1))
Step number of race.

"instance" (integer(1))
Identifies resampling instances across races and steps.

"configuration" (integer(1))
Identifies configurations across races and steps.

Result

The tuning result (instance$result) is the best performing elite of the final race. The reported performance is the average performance estimated on all used instances.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerIrace which can be applied on any black box optimization problem. See also the documentation of bbotk.

Resources

book section on tuners.

mlr3hyperband extension package for the Hyperband algorithm.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerIrace

Methods

Public methods

TunerIrace$new()

TunerIrace$optimize()

TunerIrace$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerIrace$new()

Method optimize()

Performs the tuning on a TuningInstanceSingleCrit until termination. The single evaluations and the final results will be written into the ArchiveTuning that resides in the TuningInstanceSingleCrit. The final result is returned.

Usage

TunerIrace$optimize(inst)

Arguments

inst

(TuningInstanceSingleCrit).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerIrace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# retrieve task
task = tsk("pima")

# load learner and set search space
learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE))
# \donttest{
# hyperparameter tuning on the pima indians diabetes data set
instance = tune(
  method = "irace",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  term_evals = 42
)
#> # 2022-12-21 12:17:16 UTC: Initialization
#> # Elitist race
#> # Elitist new instances: 1
#> # Elitist limit: 2
#> # nbIterations: 2
#> # minNbSurvival: 2
#> # nbParameters: 1
#> # seed: 1330381404
#> # confidence level: 0.95
#> # budget: 42
#> # mu: 5
#> # deterministic: FALSE
#> 
#> # 2022-12-21 12:17:16 UTC: Iteration 1 of 2
#> # experimentsUsedSoFar: 0
#> # remainingBudget: 42
#> # currentBudget: 21
#> # nbConfigurations: 3
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          1|          3|          1|    0.2460937500|          3|00:00:00|   NA|  NA|    NA|
#> |x|          2|          3|          1|    0.2578125000|          6|00:00:00|-1.00|0.00|1.3333|
#> |x|          3|          3|          1|    0.2617187500|          9|00:00:00|-0.33|0.11|0.8889|
#> |x|          4|          3|          1|    0.2617187500|         12|00:00:00|-0.33|0.00|0.8889|
#> |=|          5|          3|          1|    0.2609375000|         15|00:00:00|-0.20|0.04|0.8000|
#> |=|          6|          3|          1|    0.2604166667|         18|00:00:00|-0.07|0.11|0.7111|
#> |=|          7|          3|          1|    0.2611607143|         21|00:00:00|-0.14|0.02|0.7619|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           1    mean value:     0.2611607143
#> Description of the best-so-far configuration:
#>   .ID.                cp .PARENT.
#> 1    1 -3.77079679689563       NA
#> 
#> # 2022-12-21 12:17:18 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  cp
#> 1 -3.77079679689563
#> 2 -3.84314588595249
#> # 2022-12-21 12:17:18 UTC: Iteration 2 of 2
#> # experimentsUsedSoFar: 21
#> # remainingBudget: 21
#> # currentBudget: 21
#> # nbConfigurations: 4
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          8|          4|          1|    0.2812500000|          4|00:00:00|   NA|  NA|    NA|
#> |x|          2|          4|          4|    0.2734375000|          6|00:00:00|-0.57|0.21|1.1830|
#> |x|          6|          4|          4|    0.2669270833|          8|00:00:00|+0.09|0.39|0.7371|
#> |x|          1|          4|          4|    0.2617187500|         10|00:00:00|+0.23|0.42|0.5704|
#> |=|          7|          4|          4|    0.2593750000|         12|00:00:00|+0.14|0.31|0.6789|
#> |=|          3|          4|          4|    0.2610677083|         14|00:00:00|+0.19|0.32|0.6122|
#> |=|          4|          4|          4|    0.2606026786|         16|00:00:00|+0.16|0.28|0.6539|
#> |=|          5|          4|          4|    0.2602539062|         18|00:00:00|+0.06|0.18|0.7377|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:           4    mean value:     0.2602539062
#> Description of the best-so-far configuration:
#>   .ID.                cp .PARENT.
#> 4    4 -4.16510890857167        1
#> 
#> # 2022-12-21 12:17:19 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>                  cp
#> 4 -4.16510890857167
#> 5 -4.71979050208676
#> # 2022-12-21 12:17:19 UTC: Stopped because there is not enough budget left to race more than the minimum (2)
#> # You may either increase the budget or set 'minNbSurvival' to a lower value
#> # Iteration: 3
#> # nbIterations: 3
#> # experimentsUsedSoFar: 39
#> # timeUsed: 0
#> # remainingBudget: 3
#> # currentBudget: 3
#> # number of elites: 2
#> # nbConfigurations: 2
#> # Total CPU user time: 2.639, CPU sys time: 0.024, Wall-clock time: 2.664

# best performing hyperparameter configuration
instance$result
#>           cp configuration learner_param_vals  x_domain classif.ce
#> 1: -4.165109             4          <list[2]> <list[1]>  0.2602539

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp classif.ce  x_domain_cp runtime_learners           timestamp
#>  1: -3.770797  0.2460938 0.0230337028            0.012 2022-12-21 12:17:17
#>  2: -3.843146  0.2460938 0.0214260912            0.013 2022-12-21 12:17:17
#>  3: -8.530634  0.2695312 0.0001973297            0.013 2022-12-21 12:17:17
#>  4: -3.770797  0.2695312 0.0230337028            0.013 2022-12-21 12:17:17
#>  5: -3.843146  0.2695312 0.0214260912            0.013 2022-12-21 12:17:17
#>  6: -8.530634  0.2656250 0.0001973297            0.012 2022-12-21 12:17:17
#>  7: -3.770797  0.2695312 0.0230337028            0.013 2022-12-21 12:17:17
#>  8: -3.843146  0.2695312 0.0214260912            0.013 2022-12-21 12:17:17
#>  9: -8.530634  0.3085938 0.0001973297            0.013 2022-12-21 12:17:17
#> 10: -3.770797  0.2617188 0.0230337028            0.013 2022-12-21 12:17:17
#> 11: -3.843146  0.2617188 0.0214260912            0.012 2022-12-21 12:17:17
#> 12: -8.530634  0.2578125 0.0001973297            0.035 2022-12-21 12:17:17
#> 13: -3.770797  0.2578125 0.0230337028            0.013 2022-12-21 12:17:17
#> 14: -3.843146  0.2578125 0.0214260912            0.012 2022-12-21 12:17:17
#> 15: -8.530634  0.2773438 0.0001973297            0.012 2022-12-21 12:17:17
#> 16: -3.770797  0.2578125 0.0230337028            0.013 2022-12-21 12:17:18
#> 17: -3.843146  0.2578125 0.0214260912            0.012 2022-12-21 12:17:18
#> 18: -8.530634  0.3007812 0.0001973297            0.013 2022-12-21 12:17:18
#> 19: -3.770797  0.2656250 0.0230337028            0.013 2022-12-21 12:17:18
#> 20: -3.843146  0.2656250 0.0214260912            0.013 2022-12-21 12:17:18
#> 21: -8.530634  0.2460938 0.0001973297            0.013 2022-12-21 12:17:18
#> 22: -3.770797  0.2812500 0.0230337028            0.012 2022-12-21 12:17:18
#> 23: -3.843146  0.2812500 0.0214260912            0.012 2022-12-21 12:17:18
#> 24: -4.165109  0.2812500 0.0155280237            0.011 2022-12-21 12:17:18
#> 25: -4.719791  0.2851562 0.0089170465            0.012 2022-12-21 12:17:18
#> 26: -4.165109  0.2656250 0.0155280237            0.013 2022-12-21 12:17:18
#> 27: -4.719791  0.2656250 0.0089170465            0.014 2022-12-21 12:17:18
#> 28: -4.165109  0.2539062 0.0155280237            0.013 2022-12-21 12:17:18
#> 29: -4.719791  0.2695312 0.0089170465            0.012 2022-12-21 12:17:18
#> 30: -4.165109  0.2460938 0.0155280237            0.013 2022-12-21 12:17:18
#> 31: -4.719791  0.2500000 0.0089170465            0.013 2022-12-21 12:17:18
#> 32: -4.165109  0.2500000 0.0155280237            0.036 2022-12-21 12:17:19
#> 33: -4.719791  0.2500000 0.0089170465            0.014 2022-12-21 12:17:19
#> 34: -4.165109  0.2695312 0.0155280237            0.013 2022-12-21 12:17:19
#> 35: -4.719791  0.2812500 0.0089170465            0.012 2022-12-21 12:17:19
#> 36: -4.165109  0.2578125 0.0155280237            0.012 2022-12-21 12:17:19
#> 37: -4.719791  0.2578125 0.0089170465            0.012 2022-12-21 12:17:19
#> 38: -4.165109  0.2578125 0.0155280237            0.013 2022-12-21 12:17:19
#> 39: -4.719791  0.2421875 0.0089170465            0.017 2022-12-21 12:17:19
#>            cp classif.ce  x_domain_cp runtime_learners           timestamp
#>     batch_nr race step instance configuration warnings errors
#>  1:        1    1    1        9             1        0      0
#>  2:        1    1    1        9             2        0      0
#>  3:        1    1    1        9             3        0      0
#>  4:        2    1    1       10             1        0      0
#>  5:        2    1    1       10             2        0      0
#>  6:        2    1    1       10             3        0      0
#>  7:        3    1    1        8             1        0      0
#>  8:        3    1    1        8             2        0      0
#>  9:        3    1    1        8             3        0      0
#> 10:        4    1    1        1             1        0      0
#> 11:        4    1    1        1             2        0      0
#> 12:        4    1    1        1             3        0      0
#> 13:        5    1    1        2             1        0      0
#> 14:        5    1    1        2             2        0      0
#> 15:        5    1    1        2             3        0      0
#> 16:        6    1    1        3             1        0      0
#> 17:        6    1    1        3             2        0      0
#> 18:        6    1    1        3             3        0      0
#> 19:        7    1    1        4             1        0      0
#> 20:        7    1    1        4             2        0      0
#> 21:        7    1    1        4             3        0      0
#> 22:        8    2    1        7             1        0      0
#> 23:        8    2    1        7             2        0      0
#> 24:        8    2    1        7             4        0      0
#> 25:        8    2    1        7             5        0      0
#> 26:        9    2    1       10             4        0      0
#> 27:        9    2    1       10             5        0      0
#> 28:       10    2    1        3             4        0      0
#> 29:       10    2    1        3             5        0      0
#> 30:       11    2    1        9             4        0      0
#> 31:       11    2    1        9             5        0      0
#> 32:       12    2    1        4             4        0      0
#> 33:       12    2    1        4             5        0      0
#> 34:       13    2    1        8             4        0      0
#> 35:       13    2    1        8             5        0      0
#> 36:       14    2    1        1             4        0      0
#> 37:       14    2    1        1             5        0      0
#> 38:       15    2    1        2             4        0      0
#> 39:       15    2    1        2             5        0      0
#>     batch_nr race step instance configuration warnings errors
#>          resample_result
#>  1: <ResampleResult[21]>
#>  2: <ResampleResult[21]>
#>  3: <ResampleResult[21]>
#>  4: <ResampleResult[21]>
#>  5: <ResampleResult[21]>
#>  6: <ResampleResult[21]>
#>  7: <ResampleResult[21]>
#>  8: <ResampleResult[21]>
#>  9: <ResampleResult[21]>
#> 10: <ResampleResult[21]>
#> 11: <ResampleResult[21]>
#> 12: <ResampleResult[21]>
#> 13: <ResampleResult[21]>
#> 14: <ResampleResult[21]>
#> 15: <ResampleResult[21]>
#> 16: <ResampleResult[21]>
#> 17: <ResampleResult[21]>
#> 18: <ResampleResult[21]>
#> 19: <ResampleResult[21]>
#> 20: <ResampleResult[21]>
#> 21: <ResampleResult[21]>
#> 22: <ResampleResult[21]>
#> 23: <ResampleResult[21]>
#> 24: <ResampleResult[21]>
#> 25: <ResampleResult[21]>
#> 26: <ResampleResult[21]>
#> 27: <ResampleResult[21]>
#> 28: <ResampleResult[21]>
#> 29: <ResampleResult[21]>
#> 30: <ResampleResult[21]>
#> 31: <ResampleResult[21]>
#> 32: <ResampleResult[21]>
#> 33: <ResampleResult[21]>
#> 34: <ResampleResult[21]>
#> 35: <ResampleResult[21]>
#> 36: <ResampleResult[21]>
#> 37: <ResampleResult[21]>
#> 38: <ResampleResult[21]>
#> 39: <ResampleResult[21]>
#>          resample_result

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(task)
# }