Hyperparameter Tuning with Non-linear Optimization

Subclass for non-linear optimization (NLopt). Calls nloptr::nloptr from package nloptr.

Source

Johnson, G S (2020). “The NLopt nonlinear-optimization package.” https://github.com/stevengj/nlopt.

Details

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the bbotk::Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("nloptr")

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerBatchNLoptr which can be applied on any black box optimization problem. See also the documentation of bbotk.

Parameters

algorithm

character(1)
Algorithm to use. See nloptr::nloptr.print.options() for available algorithms.

x0

numeric()
Initial parameter values. Use start_values parameter to create "random" or "center" start values.

start_values

character(1)
Create "random" start values or based on "center" of search space? In the latter case, it is the center of the parameters before a trafo is applied. Custom start values can be passed via the x0 parameter.

approximate_eval_grad_f

logical(1)
Should gradients be numerically approximated via finite differences (nloptr::nl.grad). Only required for certain algorithms. Note that function evaluations required for the numerical gradient approximation will be logged as usual and are not treated differently than regular function evaluations by, e.g., Terminators.

For the meaning of other control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

Resources

There are several sections about hyperparameter optimization in the mlr3book.

Getting started with hyperparameter optimization.

An overview of all tuners can be found on our website.

Tune a support vector machine on the Sonar data set.

Learn about tuning spaces.

Estimate the model performance with nested resampling.

Learn about multi-objective optimization.

Simultaneously optimize hyperparameters and use early stopping with XGBoost.

Automate the tuning.

The gallery features a collection of case studies and demos about optimization.

Learn more advanced methods with the Practical Tuning Series.

Learn about hotstarting models.

Run the default hyperparameter configuration of learners as a baseline.

Use the Hyperband optimizer with different budget parameters.

The cheatsheet summarizes the most important functions of mlr3tuning.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerBatch -> mlr3tuning::TunerBatchFromOptimizerBatch -> TunerBatchNLoptr

Methods

Public methods

TunerBatchNLoptr$new()

TunerBatchNLoptr$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerBatchFromOptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerBatchNLoptr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerBatchNLoptr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization
# \donttest{

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  tuner = tnr("nloptr", algorithm = "NLOPT_LN_BOBYQA"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce")
)

# best performing hyperparameter configuration
instance$result
#>           cp learner_param_vals  x_domain classif.ce
#>        <num>             <list>    <list>      <num>
#> 1: -3.792442          <list[2]> <list[1]> 0.03478261

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>             cp classif.ce runtime_learners           timestamp warnings errors
#>          <num>      <num>            <num>              <POSc>    <int>  <int>
#>   1: -3.792442 0.03478261            0.006 2025-11-26 09:16:46        0      0
#>   2: -3.792442 0.03478261            0.006 2025-11-26 09:16:46        0      0
#>   3: -3.792442 0.03478261            0.006 2025-11-26 09:16:46        0      0
#>   4: -2.675049 0.04347826            0.006 2025-11-26 09:16:46        0      0
#>   5: -4.909835 0.03478261            0.026 2025-11-26 09:16:46        0      0
#>  ---                                                                          
#> 256: -3.792442 0.03478261            0.006 2025-11-26 09:17:02        0      0
#> 257: -3.792442 0.03478261            0.006 2025-11-26 09:17:02        0      0
#> 258: -3.792442 0.03478261            0.005 2025-11-26 09:17:02        0      0
#> 259: -3.792442 0.03478261            0.006 2025-11-26 09:17:02        0      0
#> 260: -3.792442 0.03478261            0.006 2025-11-26 09:17:02        0      0
#>       x_domain batch_nr  resample_result
#>         <list>    <int>           <list>
#>   1: <list[1]>        1 <ResampleResult>
#>   2: <list[1]>        2 <ResampleResult>
#>   3: <list[1]>        3 <ResampleResult>
#>   4: <list[1]>        4 <ResampleResult>
#>   5: <list[1]>        5 <ResampleResult>
#>  ---                                    
#> 256: <list[1]>      256 <ResampleResult>
#> 257: <list[1]>      257 <ResampleResult>
#> 258: <list[1]>      258 <ResampleResult>
#> 259: <list[1]>      259 <ResampleResult>
#> 260: <list[1]>      260 <ResampleResult>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))
# }