Hyperparameter Tuning with Non-linear Optimization

Subclass for non-linear optimization (NLopt). Calls nloptr::nloptr from package nloptr.

Source

Johnson, G S (2020). “The NLopt nonlinear-optimization package.” https://github.com/stevengj/nlopt.

Details

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the bbotk::Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Dictionary

This Tuner can be instantiated with the associated sugar function tnr():

tnr("nloptr")

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Optimizer

This Tuner is based on bbotk::OptimizerBatchNLoptr which can be applied on any black box optimization problem. See also the documentation of bbotk.

Parameters

algorithm

character(1)

eval_g_ineq

function()

xtol_rel

numeric(1)

xtol_abs

numeric(1)

ftol_rel

numeric(1)

ftol_abs

numeric(1)

start_values

character(1)
Create random start values or based on center of search space? In the latter case, it is the center of the parameters before a trafo is applied.

For the meaning of the control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

Resources

There are several sections about hyperparameter optimization in the mlr3book.

An overview of all tuners can be found on our website.

Learn more about tuners.

The gallery features a collection of case studies and demos about optimization.

Use the Hyperband optimizer with different budget parameters.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerBatch -> mlr3tuning::TunerBatchFromOptimizerBatch -> TunerBatchNLoptr

Methods

Public methods

TunerBatchNLoptr$new()

TunerBatchNLoptr$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerBatchFromOptimizerBatch$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerBatchNLoptr$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerBatchNLoptr$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Hyperparameter Optimization
# \donttest{

# load learner and set search space
learner = lrn("classif.rpart",
  cp = to_tune(1e-04, 1e-1, logscale = TRUE)
)

# run hyperparameter tuning on the Palmer Penguins data set
instance = tune(
  tuner = tnr("nloptr", algorithm = "NLOPT_LN_BOBYQA"),
  task = tsk("penguins"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce")
)

# best performing hyperparameter configuration
instance$result
#>           cp learner_param_vals  x_domain classif.ce
#>        <num>             <list>    <list>      <num>
#> 1: -6.698212          <list[2]> <list[1]> 0.06086957

# all evaluated hyperparameter configuration
as.data.table(instance$archive)
#>            cp classif.ce  x_domain_cp runtime_learners           timestamp
#>         <num>      <num>        <num>            <num>              <POSc>
#>  1: -6.698212 0.06086957 0.0012331145            0.031 2024-06-30 09:46:30
#>  2: -6.698212 0.06086957 0.0012331145            0.010 2024-06-30 09:46:30
#>  3: -6.698212 0.06086957 0.0012331145            0.010 2024-06-30 09:46:30
#>  4: -4.971273 0.06086957 0.0069343124            0.010 2024-06-30 09:46:31
#>  5: -8.425151 0.06086957 0.0002192822            0.010 2024-06-30 09:46:31
#>  6: -6.680943 0.06086957 0.0012545946            0.010 2024-06-30 09:46:31
#>  7: -6.715482 0.06086957 0.0012120022            0.009 2024-06-30 09:46:31
#>  8: -6.696485 0.06086957 0.0012352458            0.010 2024-06-30 09:46:31
#>  9: -6.699939 0.06086957 0.0012309868            0.010 2024-06-30 09:46:31
#> 10: -6.698212 0.06086957 0.0012331145            0.010 2024-06-30 09:46:31
#>     batch_nr warnings errors  resample_result
#>        <int>    <int>  <int>           <list>
#>  1:        1        0      0 <ResampleResult>
#>  2:        2        0      0 <ResampleResult>
#>  3:        3        0      0 <ResampleResult>
#>  4:        4        0      0 <ResampleResult>
#>  5:        5        0      0 <ResampleResult>
#>  6:        6        0      0 <ResampleResult>
#>  7:        7        0      0 <ResampleResult>
#>  8:        8        0      0 <ResampleResult>
#>  9:        9        0      0 <ResampleResult>
#> 10:       10        0      0 <ResampleResult>

# fit final model on complete data set
learner$param_set$values = instance$result_learner_param_vals
learner$train(tsk("penguins"))
# }