TunerRandomSearch

Subclass for random search tuning.

The random points are sampled by paradox::generate_design_random().

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria.

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

mlr_tuners$get("random_search")
tnr("random_search")

Parameters

batch_size (integer(1))
Maximum number of configurations to try in a batch.

See also

Super class

mlr3tuning::Tuner -> TunerRandomSearch

Methods

Public methods

TunerRandomSearch$new()

TunerRandomSearch$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

mlr3tuning::Tuner$tune()

Method new()

Creates a new instance of this R6 class.

Usage

TunerRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# see ?Tuner