TunerRandomSearch

Subclass for random search tuning.

The random points are sampled by paradox::generate_design_random().

Source

Bergstra J, Bengio Y (2012). “Random Search for Hyper-Parameter Optimization.” Journal of Machine Learning Research, 13(10), 281--305. https://jmlr.csail.mit.edu/papers/v13/bergstra12a.html.

Dictionary

This Tuner can be instantiated via the dictionary mlr_tuners or with the associated sugar function tnr():

mlr_tuners$get("random_search")
tnr("random_search")

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking of termination criteria. A batch contains of batch_size times resampling$iters jobs. E.g., if you set a batch size of 10 points and do a 5-fold cross validation, you can utilize up to 50 cores.

Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

All Tuners use a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Parameters

algorithm

character(1)

x0

numeric()

eval_g_ineq

function()

xtol_rel

numeric(1)

xtol_abs

numeric(1)

ftol_rel

numeric(1)

ftol_abs

numeric(1)

For the meaning of the control parameters, see nloptr::nloptr() and nloptr::nloptr.print.options().

The termination conditions stopval, maxtime and maxeval of nloptr::nloptr() are deactivated and replaced by the Terminator subclasses. The x and function value tolerance termination conditions (xtol_rel = 10^-4, xtol_abs = rep(0.0, length(x0)), ftol_rel = 0.0 and ftol_abs = 0.0) are still available and implemented with their package defaults. To deactivate these conditions, set them to -1.

See also

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerRandomSearch

Methods

Public methods

TunerRandomSearch$new()

TunerRandomSearch$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class.

Usage

TunerRandomSearch$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerRandomSearch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3)
library(paradox)
search_space = ParamSet$new(list(
  ParamDbl$new("cp", lower = 0.001, upper = 0.1)
))
terminator = trm("evals", n_evals = 3)
instance = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  search_space = search_space,
  terminator = terminator
)
tt = tnr("random_search")

# modifies the instance by reference
tt$optimize(instance)

#>            cp learner_param_vals  x_domain classif.ce
#> 1: 0.02752292          <list[2]> <list[1]>       0.06

# returns best configuration and best performance
instance$result

#>            cp learner_param_vals  x_domain classif.ce
#> 1: 0.02752292          <list[2]> <list[1]>       0.06

# allows access of data.table of full path of all evaluations
instance$archive

#> <ArchiveTuning>
#>            cp classif.ce                                uhash  x_domain
#> 1: 0.02752292       0.06 fda7e4c4-ed65-48e3-9c9e-f9ec50a15f5a <list[1]>
#> 2: 0.02205542       0.06 67cb463a-eb73-4b13-b8d3-beae3ad81709 <list[1]>
#> 3: 0.07970100       0.06 92371773-92f2-470d-b29f-d2107e691e69 <list[1]>
#>              timestamp batch_nr
#> 1: 2020-12-22 04:45:59        1
#> 2: 2020-12-22 04:46:00        2
#> 3: 2020-12-22 04:46:00        3